refactor(shares): split service.go by concern
service.go was 3348 lines mixing block-store construction and wrapping,
share lifecycle, per-field share-config CRUD, test hooks, and block-store
operations/stats. Each moves to its own file: blockstore_config.go,
lifecycle.go, share_config.go, testing.go and blockstore_ops.go, leaving
service.go with the types, the Service struct, and the share lookup and
notification surface.

Pure code motion — every non-import line is preserved verbatim.
