+++ pkg/controlplane/runtime/shares/service.go
@@ -2,32 +2,17 @@ package shares
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"math"
 	"net/url"
-	"os"
-	"path/filepath"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/marmos91/dittofs/internal/logger"
-	"github.com/marmos91/dittofs/internal/pathutil"
 	"github.com/marmos91/dittofs/pkg/block"
-	"github.com/marmos91/dittofs/pkg/block/chunker"
-	"github.com/marmos91/dittofs/pkg/block/compression"
-	"github.com/marmos91/dittofs/pkg/block/encryption"
-	"github.com/marmos91/dittofs/pkg/block/encryption/keyprovider"
 	"github.com/marmos91/dittofs/pkg/block/engine"
 	"github.com/marmos91/dittofs/pkg/block/local"
-	"github.com/marmos91/dittofs/pkg/block/local/fs"
-	localmemory "github.com/marmos91/dittofs/pkg/block/local/memory"
-	"github.com/marmos91/dittofs/pkg/block/remote"
-	remotememory "github.com/marmos91/dittofs/pkg/block/remote/memory"
-	remotes3 "github.com/marmos91/dittofs/pkg/block/remote/s3"
 	"github.com/marmos91/dittofs/pkg/controlplane/models"
 	"github.com/marmos91/dittofs/pkg/metadata"
 )
@@ -282,50 +267,6 @@ type MetadataDurableExtentSetter interface {
 	SetDurableExtentResolver(fn metadata.DurableExtentFunc)
 }
 
-// BlockStoreConfigProvider resolves block store configurations from the control plane DB.
-//
-// A share's LocalBlockStoreID/RemoteBlockStoreID normally hold the block store
-// row's UUID, but the REST UpdateShare path historically persisted the raw
-// *name* instead (#1312). Resolution therefore tries the UUID first and falls
-// back to a kind-scoped name lookup so shares whose rows hold a name still load
-// after a restart.
-type BlockStoreConfigProvider interface {
-	GetBlockStoreByID(ctx context.Context, id string) (*models.BlockStoreConfig, error)
-	GetBlockStore(ctx context.Context, name string, kind models.BlockStoreKind) (*models.BlockStoreConfig, error)
-}
-
-// resolveBlockStoreConfig resolves a block store reference that may be either a
-// UUID (the canonical form) or a name (#1312 legacy rows). It tries the UUID
-// lookup first; on not-found it falls back to a name lookup scoped to the
-// expected kind. The kind scope keeps a local name from accidentally resolving
-// to a same-named remote store and vice versa.
-func resolveBlockStoreConfig(
-	ctx context.Context,
-	provider BlockStoreConfigProvider,
-	ref string,
-	kind models.BlockStoreKind,
-) (*models.BlockStoreConfig, error) {
-	cfg, err := provider.GetBlockStoreByID(ctx, ref)
-	if err == nil {
-		return cfg, nil
-	}
-	if !errors.Is(err, models.ErrStoreNotFound) {
-		return nil, err
-	}
-	// Fall back to name resolution for legacy rows that stored the name.
-	byName, nameErr := provider.GetBlockStore(ctx, ref, kind)
-	if nameErr != nil {
-		// A real operational error (DB/context) on the name path must not be
-		// masked as "not found"; only collapse to the familiar ID-lookup error
-		// when the name is genuinely absent too.
-		if errors.Is(nameErr, models.ErrStoreNotFound) {
-			return nil, err
-		}
-		return nil, nameErr
-	}
-	return byName, nil
-}
-
 // ShareStore is the narrow subset of pkg/controlplane/store.ShareStore that
 // DisableShare / EnableShare need. Defined here so callers can pass any store
 // that satisfies it (the concrete GORMStore does) without importing the
@@ -333,146 +274,6 @@ func resolveBlockStoreConfig(
 type ShareStore interface {
 	GetShare(ctx context.Context, name string) (*models.Share, error)
 	UpdateShare(ctx context.Context, share *models.Share) error
-}
-
-// LocalStoreDefaults holds default sizing for per-share local stores.
-type LocalStoreDefaults struct {
-	MaxSize uint64 // Maximum local store size per share (0 = unlimited)
-
-	// ReadBufferBytes is the per-share read buffer budget in bytes (0 = disabled).
-	ReadBufferBytes int64
-
-	// MaxLogBytes is the effective append-log pressure budget (bytes) applied
-	// to a share when its per-share block store config does NOT carry an
-	// explicit `max_log_bytes`. It resolves the global config
-	// blockstore.local.max_log_bytes (when set) or, failing that, the
-	// system-deduced default (DeduceDefaults.MaxLogBytes). 0 leaves the
-	// FSStore's own internal default in force. Precedence at the share level
-	// is: per-store config["max_log_bytes"] > this global/deduced default.
-	MaxLogBytes int64
-
-	// DefaultRemoteCacheSize is the on-disk ceiling applied to a share's
-	// local tier when a REMOTE block store is configured but no explicit
-	// per-share LocalStoreSize is set. With a remote configured the local
-	// tier is a bounded write-through cache; without this ceiling a fast
-	// writer could exhaust the host volume. 0 leaves the conditional ceiling
-	// off (the share keeps its system-deduced local size even with a
-	// remote). Local-only shares ignore this entirely.
-	DefaultRemoteCacheSize uint64
-
-	// BackpressureMaxWait is how long a write stalls waiting for the syncer
-	// to drain (freeing cache space) before returning ErrDiskFull, when the
-	// remote is healthy but every cached chunk is unsynced. 0 defers to the
-	// FSStore default (60s).
-	BackpressureMaxWait time.Duration
-}
-
-// SyncerDefaults holds default syncer configuration applied to all shares.
-type SyncerDefaults struct {
-	ParallelDownloads  int
-	PrefetchBlocks     int
-	SmallFileThreshold int64
-	UploadInterval     time.Duration
-	UploadDelay        time.Duration
-
-	// PrefetchWorkers is the number of read buffer prefetch workers per share (0 = disabled).
-	PrefetchWorkers int
-}
-
-// sharedRemote holds a reference-counted remote store shared across shares.
-type sharedRemote struct {
-	store    remote.RemoteStore
-	refCount int
-}
-
-// nonClosingRemote wraps a remote.RemoteStore and makes Close() a no-op.
-// This prevents engine.Store.Close() from closing the shared remote;
-// the shares.Service.releaseRemoteStore handles actual closing via ref counting.
-type nonClosingRemote struct {
-	remote.RemoteStore
-}
-
-func (n *nonClosingRemote) Close() error { return nil }
-
-// Durable delegates to the wrapped remote chain so the no-op-Close wrapper
-// still implements block.DurabilityReporter. Without this, embedding
-// remote.RemoteStore (which has no Durable method) would silently drop the
-// capability, and engine.Store.RemoteDurable() would type-assert to
-// block.DurabilityReporter, fail, and report NOT durable for every production
-// S3-remote share — breaking the honest COMMIT/CLOSE contract (#1274) with a
-// spurious ErrNotDurableYet on every commit.
-func (n *nonClosingRemote) Durable() bool { return block.IsDurable(n.RemoteStore) }
-
-// --- remote.RemoteBlockStore proxy (#1414 object packing) ---
-//
-// The no-op-Close wrapper embeds only remote.RemoteStore, so without these
-// forwards it would silently HIDE the block-keyed surface (PutBlock/GetBlock/
-// GetBlockRange/DeleteBlock/WalkBlocks) of a wrapped store that implements it —
-// exactly the trap the Durable() forward above fixes for durability. The
-// snapshot durability-verify gate reaches the block store via
-// engine.Store.RemoteStore() (this wrapper), so it MUST be able to probe a
-// packed blocks/<id> object. Each method delegates to the embedded store when it
-// implements RemoteBlockStore; otherwise it returns remote.ErrChunkReadUnsupported
-// (a standalone-only remote is never asked for a block via the locator path).
-func (n *nonClosingRemote) blockInner() (remote.RemoteBlockStore, error) {
-	if rbs, ok := n.RemoteStore.(remote.RemoteBlockStore); ok {
-		return rbs, nil
-	}
-	return nil, remote.ErrChunkReadUnsupported
-}
-
-func (n *nonClosingRemote) PutBlock(ctx context.Context, blockID string, r io.Reader) error {
-	rbs, err := n.blockInner()
-	if err != nil {
-		return err
-	}
-	return rbs.PutBlock(ctx, blockID, r)
-}
-
-func (n *nonClosingRemote) GetBlock(ctx context.Context, blockID string) ([]byte, error) {
-	rbs, err := n.blockInner()
-	if err != nil {
-		return nil, err
-	}
-	return rbs.GetBlock(ctx, blockID)
-}
-
-func (n *nonClosingRemote) GetBlockRange(ctx context.Context, blockID string, offset, length int64) ([]byte, error) {
-	rbs, err := n.blockInner()
-	if err != nil {
-		return nil, err
-	}
-	return rbs.GetBlockRange(ctx, blockID, offset, length)
-}
-
-func (n *nonClosingRemote) DeleteBlock(ctx context.Context, blockID string) error {
-	rbs, err := n.blockInner()
-	if err != nil {
-		return err
-	}
-	return rbs.DeleteBlock(ctx, blockID)
-}
-
-func (n *nonClosingRemote) WalkBlocks(ctx context.Context, fn func(blockID string, meta block.Meta) error) error {
-	rbs, err := n.blockInner()
-	if err != nil {
-		return err
-	}
-	return rbs.WalkBlocks(ctx, fn)
-}
-
-// ReadChunk delegates the remote.ChunkReader capability (#1414) to the wrapped
-// store. The syncer's read path type-asserts ChunkReader on ITS remote — this
-// wrapper — to serve a chunk whose only copy lives inside a packed block.
-// Without this forward every cold read of a packed chunk (local copy lost:
-// restart, eviction, torn tail) on a production share failed with
-// ErrChunkReadUnsupported instead of recovering from the remote block.
-func (n *nonClosingRemote) ReadChunk(ctx context.Context, blockID string, offset, length int64, hash block.ContentHash) ([]byte, error) {
-	cr, ok := n.RemoteStore.(remote.ChunkReader)
-	if !ok {
-		return nil, remote.ErrChunkReadUnsupported
-	}
-	return cr.ReadChunk(ctx, blockID, offset, length, hash)
 }
 
 // Service manages share registration, lookup, and configuration.
@@ -555,1528 +356,6 @@ func sanitizeShareName(name string) string {
 	name = strings.TrimPrefix(name, "/")
 	return url.PathEscape(name)
 }
-
-// buildSyncerConfigFromDefaults merges SyncerDefaults into a engine.SyncerConfig.
-func buildSyncerConfigFromDefaults(defaults *SyncerDefaults) engine.SyncerConfig {
-	cfg := engine.DefaultConfig()
-	if defaults == nil {
-		return cfg
-	}
-	if defaults.ParallelDownloads > 0 {
-		cfg.ParallelDownloads = defaults.ParallelDownloads
-	}
-	if defaults.PrefetchBlocks > 0 {
-		cfg.PrefetchBlocks = defaults.PrefetchBlocks
-	}
-	if defaults.SmallFileThreshold != 0 {
-		cfg.SmallFileThreshold = defaults.SmallFileThreshold
-	}
-	if defaults.UploadInterval > 0 {
-		cfg.UploadInterval = defaults.UploadInterval
-	}
-	if defaults.UploadDelay > 0 {
-		cfg.UploadDelay = defaults.UploadDelay
-	}
-	return cfg
-}
-
-// remotePinnedUploads returns the per-remote parallel_uploads override: > 0
-// pins the carver's upload window, 0/absent keeps the adaptive auto-tune
-// (#1407 / #1432). Any lookup/parse miss falls back to 0 (adaptive) so a
-// malformed value never blocks share creation — validateParallelUploads already
-// rejects out-of-range values at store-create time. The value arrives as a JSON
-// number (float64) from the stored config blob. We still re-validate here
-// (integer-ness + 0..engine.MaxParallelUploads clamp) as defense-in-depth against a
-// corrupted or old-server config blob that skipped validateParallelUploads —
-// silently truncating 2.5 or honoring an absurd window would risk FD/goroutine
-// exhaustion.
-func remotePinnedUploads(ctx context.Context, provider BlockStoreConfigProvider, ref string) int {
-	if ref == "" || provider == nil {
-		return 0
-	}
-	cfg, err := resolveBlockStoreConfig(ctx, provider, ref, models.BlockStoreKindRemote)
-	if err != nil {
-		return 0
-	}
-	m, err := cfg.GetConfig()
-	if err != nil {
-		return 0
-	}
-	var n int
-	switch v := m["parallel_uploads"].(type) {
-	case float64:
-		if v != math.Trunc(v) { // fractional => malformed, fall back to adaptive
-			return 0
-		}
-		n = int(v)
-	case int:
-		n = v
-	default:
-		return 0
-	}
-	if n <= 0 || n > engine.MaxParallelUploads {
-		return 0
-	}
-	return n
-}
-
-func (s *Service) AddShare(
-	ctx context.Context,
-	config *ShareConfig,
-	storeProvider MetadataStoreProvider,
-	metadataSvc MetadataServiceRegistrar,
-	blockStoreProvider BlockStoreConfigProvider,
-	localStoreDefaults *LocalStoreDefaults,
-	syncerDefaults *SyncerDefaults,
-) error {
-	if config.Name == "" {
-		return errors.New("cannot add share with empty name")
-	}
-	// File handles are encoded as "<shareName>:<uuid>" and decoded by
-	// splitting on the FIRST ':'. A share name containing ':' yields handles
-	// whose UUID component fails to parse, silently bricking every file in the
-	// share. Reject it up front with an ErrInvalidArgument StoreError.
-	if strings.Contains(config.Name, ":") {
-		return metadata.NewInvalidArgumentError(
-			fmt.Sprintf("share name %q must not contain ':'", config.Name))
-	}
-
-	if config.LocalBlockStoreID != "" && blockStoreProvider == nil {
-		return fmt.Errorf("block store provider is required when LocalBlockStoreID is set for share %q", config.Name)
-	}
-
-	if metadataSvc == nil {
-		return fmt.Errorf("metadata service registrar is required for share %q", config.Name)
-	}
-
-	// Phase 0: Reserve the share name under s.mu BEFORE any side-effecting init
-	// (root-dir create, block-store start, metadata RegisterStoreForShare). This
-	// closes the AddShare(sameName) race (REVIEW M2): the OLD ordering let two
-	// racing callers both run RegisterStoreForShare (last writer wins on the
-	// MetadataService store map) and only THEN recheck the registry under the
-	// lock — so the loser would tear down its block store yet leave its metadata
-	// store registered, pointing the MetadataService at a different store than
-	// the registry exposes. Reserving here serializes on the name first: the
-	// loser fails before touching any shared store, so no mismatch can form.
-	//
-	// The reservation is NOT inserted into registry, so a half-built share is
-	// never visible to handlers; it is converted to a registry entry only in
-	// Phase 4 after all init succeeds.
-	if err := s.reserveShareName(config.Name); err != nil {
-		return err
-	}
-	// Track whether the reservation still needs releasing. Phase 4 hands the
-	// reservation off to the registry entry (under the same lock) and clears
-	// this flag; every failure path releases it.
-	reservationHeld := true
-	defer func() {
-		if reservationHeld {
-			s.releaseShareName(config.Name)
-		}
-	}()
-
-	// Phase 1: Build share struct (resolves metadata store, creates root dir).
-	// Does NOT insert into registry yet -- share is invisible to handlers.
-	share, metadataStore, err := s.prepareShare(ctx, config, storeProvider)
-	if err != nil {
-		return err
-	}
-
-	// Phase 2: Create per-share BlockStore if local block store config is provided.
-	if config.LocalBlockStoreID != "" {
-		if err := s.createBlockStoreForShare(ctx, share, config, blockStoreProvider, metadataStore, localStoreDefaults, syncerDefaults); err != nil {
-			return fmt.Errorf("failed to create block store for share %q: %w", config.Name, err)
-		}
-	}
-
-	// cleanupShare releases resources for a share that failed to fully initialize.
-	cleanupShare := func() {
-		if share.BlockStore != nil {
-			_ = share.BlockStore.Close()
-		}
-		if share.remoteConfigID != "" {
-			s.releaseRemoteStore(share.remoteConfigID)
-		}
-	}
-
-	// Phase 2.5: Reconcile metadata file sizes against the local journal's durable
-	// high-water mark (#1687 safety net). WRITE/COMMIT now commit file size via a
-	// relaxed (deferred-fsync) metadata transaction, so a crash between a relaxed
-	// size commit and its background fsync can leave metadata.Size smaller than the
-	// data actually made durable in the journal. This grows metadata.Size up to the
-	// journal size (max-only, never shrinks) BEFORE the share is registered and any
-	// protocol handler can read it, so ACK'd bytes are never truncated.
-	if config.LocalBlockStoreID != "" && share.BlockStore != nil {
-		if err := reconcileMetadataSizeFromJournal(ctx, metadataStore, share.BlockStore); err != nil {
-			cleanupShare()
-			return fmt.Errorf("failed to reconcile metadata sizes for share %q: %w", config.Name, err)
-		}
-	}
-
-	// Phase 3: Register metadata store. Safe to call unconditionally now: the
-	// Phase-0 reservation guarantees no other AddShare for this name is in
-	// flight, so this RegisterStoreForShare cannot be raced into a
-	// last-writer-wins mismatch.
-	if err := metadataSvc.RegisterStoreForShare(config.Name, metadataStore); err != nil {
-		cleanupShare()
-		return fmt.Errorf("failed to configure metadata for share: %w", err)
-	}
-
-	// Apply the per-share metadata writeback tier (#1757) parsed from the local
-	// store config in createBlockStoreForShare. Set explicitly (true or false) so
-	// a re-add that toggles writeback off is honored. Only the concrete
-	// *metadata.Service implements the setter.
-	if wb, ok := metadataSvc.(MetadataWritebackSetter); ok {
-		wb.SetShareWriteback(config.Name, share.writeback)
-	}
-
-	// Let size commits consult the block store's durable extent. The resolver is
-	// share-agnostic, so re-installing it on every AddShare is idempotent.
-	if de, ok := metadataSvc.(MetadataDurableExtentSetter); ok {
-		de.SetDurableExtentResolver(s.durableExtent)
-	}
-
-	// Phase 4: Convert the reservation into a registry entry under s.mu.
-	// Only now is the share visible to protocol handlers. The reservation has
-	// held the name exclusively since Phase 0, so registry[name] cannot already
-	// exist here; we assert it defensively and hand off the reservation.
-	s.mu.Lock()
-	if _, exists := s.registry[config.Name]; exists {
-		// Should be unreachable while the reservation is held, but stay
-		// fail-safe: tear down rather than overwrite an existing share.
-		s.mu.Unlock()
-		cleanupShare()
-		// Deregister the metadata store we just published so we do not leak a
-		// registration for a share we are refusing to finalize.
-		if remover, ok := metadataSvc.(MetadataServiceDeregistrar); ok {
-			remover.RemoveStoreForShare(config.Name)
-		}
-		return fmt.Errorf("share %q already exists", config.Name)
-	}
-	s.registry[config.Name] = share
-	if share.BlockStore != nil {
-		s.blockStoreCache.Store(config.Name, share.BlockStore)
-	}
-	delete(s.reservations, config.Name)
-	reservationHeld = false
-	s.mu.Unlock()
-
-	s.notifyShareChange()
-
-	return nil
-}
-
-// durableExtent reports how far a payload's bytes are on stable storage in its
-// share's block store. A share with no block store answers "unknown" (false),
-// and the caller then commits the size unclamped as before.
-//
-// It is the commit-time half of the size invariant whose restart-time half is
-// reconcileMetadataSizeFromJournal: a persisted size never runs ahead of the
-// durable data, and share start grows it back up to that data.
-func (s *Service) durableExtent(shareName string, payloadID metadata.PayloadID) (int64, bool) {
-	bs, err := s.GetBlockStoreForShare(shareName)
-	if err != nil || bs == nil {
-		return 0, false
-	}
-	return bs.DurableExtent(context.Background(), payloadID)
-}
-
-// reconcileMetadataSizeFromJournal grows each file's metadata size up to the
-// local journal's durable high-water mark (#1687 crash-safety net). It is the
-// load-bearing counterpart to the relaxed (deferred-fsync) size commits done on
-// WRITE/COMMIT: if the server crashed after the data was fsync'd into the journal
-// but before the relaxed metadata size fsync ran, metadata.Size is stale/smaller
-// than the durable data. Reading that file would truncate ACK'd bytes (#588
-// class). This runs on share start, before the share is visible to handlers, and
-// only ever GROWS the size — a legitimate shrink (SetAttr/truncate) commits
-// strictly and is never rolled back here.
-//
-// bs.Local.ListFiles/FileSize are served from the journal's in-memory index
-// (populated by recover() at open), so the common no-mismatch case is cheap: a
-// point read of metadata per file, and a strict PutFile only on an actual gap.
-func reconcileMetadataSizeFromJournal(ctx context.Context, metadataStore metadata.Store, bs *engine.Store) error {
-	if bs == nil {
-		return nil
-	}
-	localStore := bs.Local()
-	if localStore == nil {
-		return nil
-	}
-	for _, id := range localStore.ListFiles(ctx) {
-		journalSize, ok := localStore.FileSize(ctx, id)
-		if !ok {
-			continue
-		}
-		f, err := metadataStore.GetFileByPayloadID(ctx, metadata.PayloadID(id))
-		if err != nil {
-			if metadata.IsNotFoundError(err) {
-				// Orphan journal entry (no metadata file) — nothing to reconcile.
-				continue
-			}
-			// A real store error (I/O, corruption) must not be silently swallowed —
-			// skipping would leave metadata.Size stale and truncate reads.
-			return fmt.Errorf("reconcile size: lookup payload %s: %w", id, err)
-		}
-		if f == nil {
-			continue
-		}
-		// max-only, overflow-safe: f.Size is uint64, journalSize a non-negative
-		// offset; compare in uint64 so a huge Size can never cast negative and shrink.
-		if journalSize < 0 || f.Size >= uint64(journalSize) {
-			continue
-		}
-		// Grow to the journal high-water mark under a STRICT transaction; re-read
-		// inside the txn so a concurrent legitimate update can't be clobbered.
-		if err := metadataStore.WithTransaction(ctx, func(tx metadata.Transaction) error {
-			cur, err := tx.GetFileByPayloadID(ctx, metadata.PayloadID(id))
-			if err != nil {
-				return err
-			}
-			if cur == nil || journalSize < 0 || cur.Size >= uint64(journalSize) {
-				return nil // another writer already caught up; never shrink
-			}
-			cur.Size = uint64(journalSize)
-			return tx.PutFile(ctx, cur)
-		}); err != nil {
-			return fmt.Errorf("reconcile size for payload %s: %w", id, err)
-		}
-	}
-	return nil
-}
-
-// reserveShareName claims a share name for an in-flight AddShare. It fails if
-// the name is already registered or already reserved by a concurrent AddShare.
-func (s *Service) reserveShareName(name string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if _, exists := s.registry[name]; exists {
-		return fmt.Errorf("share %q already exists", name)
-	}
-	if _, reserved := s.reservations[name]; reserved {
-		return fmt.Errorf("share %q is already being added", name)
-	}
-	s.reservations[name] = struct{}{}
-	return nil
-}
-
-// PendingShareRegistrations reports how many AddShare calls hold a name
-// reservation they have not yet published to the registry. Such a share is
-// invisible to DistinctRemoteStores even though its block store — and the
-// background legacy migration that block store starts — is already running, so
-// callers that must reason about every share on a remote have to treat a
-// non-zero count as "the share set is not knowable right now".
-func (s *Service) PendingShareRegistrations() int {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return len(s.reservations)
-}
-
-// releaseShareName drops an in-flight AddShare reservation. Idempotent.
-func (s *Service) releaseShareName(name string) {
-	s.mu.Lock()
-	delete(s.reservations, name)
-	s.mu.Unlock()
-}
-
-// rootModeForDefaultPermission projects a share's default_permission onto the
-// share root directory's POSIX permission bits. NFSv3 has no ACL transport, so
-// for its clients the mode bits are the only access signal — and the Linux
-// client enforces them *client-side*, before any RPC. The bits can therefore
-// only restrict, never grant (the server's export gate + ACL remain the real
-// authority); but a too-narrow root silently blocked legitimate access. The
-// only case that needed opening is read-write: a non-root user on a read-write
-// share was denied at the 0755 root before reaching the server. So read-write
-// (and admin) widen "other" to rwx; read and the secure none/unset default keep
-// the historical 0755 (the export gate still denies ungranted access server-
-// side). Per-user grants remain NFSv4/SMB-only — mode bits cannot express them.
-func rootModeForDefaultPermission(perm string) uint32 {
-	switch models.SharePermission(perm) {
-	case models.PermissionReadWrite, models.PermissionAdmin:
-		return 0o777
-	default: // none / unset / read: preserve the historical 0755 root
-		return 0o755
-	}
-}
-
-// prepareShare validates config, resolves the metadata store, and creates the
-// root directory. Returns the built Share (not yet in the registry) and the
-// metadata store. The caller (AddShare) is responsible for inserting the share
-// into the registry after all initialization (including BlockStore) succeeds.
-func (s *Service) prepareShare(
-	ctx context.Context,
-	config *ShareConfig,
-	storeProvider MetadataStoreProvider,
-) (*Share, metadata.Store, error) {
-	// Early duplicate check (optimistic -- AddShare rechecks under write lock).
-	s.mu.RLock()
-	if _, exists := s.registry[config.Name]; exists {
-		s.mu.RUnlock()
-		return nil, nil, fmt.Errorf("share %q already exists", config.Name)
-	}
-	s.mu.RUnlock()
-
-	if storeProvider == nil {
-		return nil, nil, errors.New("metadata store provider not initialized")
-	}
-
-	metadataStore, err := storeProvider.GetMetadataStore(config.MetadataStore)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	rootAttr := config.RootAttr
-	if rootAttr == nil {
-		rootAttr = &metadata.FileAttr{}
-	}
-	if rootAttr.Type == 0 {
-		rootAttr.Type = metadata.FileTypeDirectory
-	}
-	if rootAttr.Mode == 0 {
-		rootAttr.Mode = rootModeForDefaultPermission(config.DefaultPermission)
-	}
-	if rootAttr.Atime.IsZero() {
-		now := time.Now()
-		rootAttr.Atime = now
-		rootAttr.Mtime = now
-		rootAttr.Ctime = now
-	}
-
-	rootFile, err := metadataStore.CreateRootDirectory(ctx, config.Name, rootAttr)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to create root directory: %w", err)
-	}
-
-	// Propagate a read-only share into the metadata store's ShareOptions.
-	// The runtime seeds metadata via CreateRootDirectory (not CreateShare),
-	// which leaves Options zero-valued, so without this the permission layer
-	// never sees ShareOptions.ReadOnly — the store-level discriminator that lets
-	// it deny writes/creates with ErrReadOnly (EROFS) rather than ErrAccessDenied
-	// (EACCES). Only touch read-only shares; read-write shares use the zero default.
-	if config.ReadOnly {
-		shareOpts, err := metadataStore.GetShareOptions(ctx, config.Name)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to read share options: %w", err)
-		}
-		shareOpts.ReadOnly = true
-		if err := metadataStore.UpdateShareOptions(ctx, config.Name, shareOpts); err != nil {
-			return nil, nil, fmt.Errorf("failed to set read-only share options: %w", err)
-		}
-	}
-
-	rootHandle, err := metadata.EncodeFileHandle(rootFile)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to encode root handle: %w", err)
-	}
-
-	allowAuthSys := config.AllowAuthSys
-	if !config.AllowAuthSysSet && !allowAuthSys {
-		allowAuthSys = true
-	}
-
-	share := &Share{
-		Name:                             config.Name,
-		MetadataStore:                    config.MetadataStore,
-		RootHandle:                       rootHandle,
-		ReadOnly:                         config.ReadOnly,
-		Enabled:                          config.Enabled,
-		EncryptData:                      config.EncryptData,
-		AclFlagInheritedCanonicalization: config.AclFlagInheritedCanonicalization,
-		AccessBasedEnumeration:           config.AccessBasedEnumeration,
-		ChangeNotifyDisabled:             config.ChangeNotifyDisabled,
-		StreamsDisabled:                  config.StreamsDisabled,
-		ContinuousAvailability:           config.ContinuousAvailability,
-		AllowMFsymlink:                   config.AllowMFsymlink,
-		TrashEnabled:                     config.TrashEnabled,
-		TrashRetentionDays:               config.TrashRetentionDays,
-		TrashRestrictToAdmin:             config.TrashRestrictToAdmin,
-		TrashMaxBytes:                    config.TrashMaxBytes,
-		TrashExcludePatterns:             config.TrashExcludePatterns,
-		DefaultPermission:                config.DefaultPermission,
-		Squash:                           config.Squash,
-		AnonymousUID:                     config.AnonymousUID,
-		AnonymousGID:                     config.AnonymousGID,
-		DisableReaddirplus:               config.DisableReaddirplus,
-		AllowAuthSys:                     allowAuthSys,
-		RequireKerberos:                  config.RequireKerberos,
-		MinKerberosLevel:                 config.MinKerberosLevel,
-		NetgroupName:                     config.NetgroupName,
-		BlockedOperations:                config.BlockedOperations,
-		RetentionPolicy:                  config.RetentionPolicy,
-		RetentionTTL:                     config.RetentionTTL,
-	}
-
-	return share, metadataStore, nil
-}
-
-// mergeLocalStoreDefaults returns a copy of the system defaults with per-share
-// overrides applied. Non-zero ShareConfig values take precedence.
-//
-// remoteConfigured signals that the share has a remote block store, which
-// makes the local tier a bounded write-through cache rather than durable
-// storage. In that case, when the operator set no explicit per-share size
-// (config.LocalStoreSize == 0), apply the DefaultRemoteCacheSize ceiling so
-// a fast writer cannot exhaust the host volume — this takes precedence over
-// the generic system-deduced MaxSize, which is sized for the durable
-// local-only tier rather than a transient cache. An explicit per-share
-// --local-store-size always wins. Local-only shares keep the existing
-// MaxSize unchanged.
-func mergeLocalStoreDefaults(defaults *LocalStoreDefaults, config *ShareConfig, remoteConfigured bool) *LocalStoreDefaults {
-	if defaults == nil {
-		defaults = &LocalStoreDefaults{}
-	}
-	merged := *defaults // shallow copy
-	switch {
-	case config.LocalStoreSize > 0:
-		// Explicit per-share override always wins.
-		merged.MaxSize = uint64(config.LocalStoreSize)
-	case remoteConfigured && merged.DefaultRemoteCacheSize > 0:
-		// Remote-backed share, no explicit override: bound the
-		// write-through cache at the remote-cache default.
-		merged.MaxSize = merged.DefaultRemoteCacheSize
-		logger.Info("applying default local-cache ceiling for remote-backed share",
-			"share", config.Name,
-			"max_size", merged.MaxSize)
-	}
-	if config.ReadBufferSize > 0 {
-		merged.ReadBufferBytes = config.ReadBufferSize
-	}
-	return &merged
-}
-
-// legacyLocalOnlyMigrator is the local-store surface the shares service drives
-// to finish an async pre-journal local-only migration in the background.
-// Implemented by the journal-backed fs store; other backends never satisfy it,
-// so the drive block is a no-op for them.
-type legacyLocalOnlyMigrator interface {
-	MigratedFromLegacyLocalOnly() bool
-	LegacyPendingPayloads() []string
-	MaterializeLegacyPayload(payloadID string) error
-	FinishLegacyMigration() error
-}
-
-// createBlockStoreForShare creates and starts a per-share BlockStore.
-func (s *Service) createBlockStoreForShare(
-	ctx context.Context,
-	share *Share,
-	config *ShareConfig,
-	blockStoreProvider BlockStoreConfigProvider,
-	fileChunkStore block.EngineFileChunkStore,
-	localStoreDefaults *LocalStoreDefaults,
-	syncerDefaults *SyncerDefaults,
-) error {
-	// Resolve local block store config from DB (by UUID, or by name for #1312
-	// legacy rows).
-	localCfg, err := resolveBlockStoreConfig(ctx, blockStoreProvider, config.LocalBlockStoreID, models.BlockStoreKindLocal)
-	if err != nil {
-		return fmt.Errorf("failed to resolve local block store config %q: %w", config.LocalBlockStoreID, err)
-	}
-	if localCfg.Kind != models.BlockStoreKindLocal {
-		return fmt.Errorf("block store config %q has kind %q, expected %q", config.LocalBlockStoreID, localCfg.Kind, models.BlockStoreKindLocal)
-	}
-
-	// Merge per-share size overrides into effective defaults. A configured
-	// remote makes the local tier a bounded write-through cache, so the
-	// conditional default ceiling applies (see mergeLocalStoreDefaults).
-	remoteConfigured := config.RemoteBlockStoreID != ""
-	effectiveDefaults := mergeLocalStoreDefaults(localStoreDefaults, config, remoteConfigured)
-
-	// A remote-backed share whose local dir still holds the pre-journal
-	// blobs/+logs/ layout is migrated in place: the local dirs are archived aside
-	// so the journal opens clean, and the bytes are re-materialized from the
-	// remote via a cold seed below. A local-only share passes false so the
-	// guardrail refuses to open a legacy dir (its bytes are the sole copy).
-	localStore, err := CreateLocalStoreFromConfig(ctx, localCfg.Type, localCfg, config.Name, effectiveDefaults, fileChunkStore, remoteConfigured)
-	if err != nil {
-		return fmt.Errorf("failed to create local store: %w", err)
-	}
-
-	var remoteStore remote.RemoteStore
-	var remoteConfigID string
-	if config.RemoteBlockStoreID != "" {
-		remoteStore, remoteConfigID, err = s.acquireRemoteStore(ctx, config.RemoteBlockStoreID, blockStoreProvider)
-		if err != nil {
-			_ = localStore.Close()
-			return fmt.Errorf("failed to create remote store: %w", err)
-		}
-	}
-
-	// Eviction requires a remote store (so evicted blocks can be re-fetched) and
-	// must not be pin mode (pin keeps blocks stored locally indefinitely).
-	localStore.SetEvictionEnabled(remoteStore != nil && config.RetentionPolicy != block.RetentionPin)
-	// Note: SetSkipFsync was removed. Local-disk durability is now
-	// unconditional (the syncer will refetch from S3 on the rare crash path).
-
-	syncerCfg := buildSyncerConfigFromDefaults(syncerDefaults)
-	// A per-remote parallel_uploads override pins the carver's upload window;
-	// 0 (the default) keeps the adaptive auto-tune (#1407 / #1432).
-	if pinned := remotePinnedUploads(ctx, blockStoreProvider, config.RemoteBlockStoreID); pinned > 0 {
-		syncerCfg.ParallelUploads = pinned
-	}
-
-	// Wrap shared remote in nonClosingRemote so engine.Close() doesn't close it;
-	// releaseRemoteStore handles actual closing via ref counting.
-	var engineRemote remote.RemoteStore
-	if remoteStore != nil {
-		engineRemote = &nonClosingRemote{remoteStore}
-	}
-
-	syncer := engine.NewSyncer(localStore, engineRemote, fileChunkStore, syncerCfg)
-
-	// Write-path backpressure is now internal to the journal-backed local store
-	// (Config.EvictMaxWait): a full local cache stalls the writer while the carve
-	// dispatcher drains unsynced bytes, so there is no SetBackpressureSource
-	// wiring to inject here anymore.
-
-	// Wire the block-carve substrate (#1414 object packing, PR3 global flip).
-	// Assert on the RAW remoteStore, not engineRemote: the carver holds its own
-	// dedicated RemoteBlockStore reference, so wiring it straight to the
-	// underlying store skips the no-op-Close wrapper's per-call forwarding hop.
-	// (The wrapper does forward the block-keyed surface + ReadChunk these days,
-	// so this is conventional rather than required.) Every shipped remote (memory,
-	// s3) implements RemoteBlockStore, so this flips carve on for EVERY share
-	// with a remote — no feature flag. SetSyncedHashStore (called by engine.New
-	// below) derives the blockCommitter from the same per-share metadata store
-	// and recomputes carveActive; once all deps are set, new writes route to the
-	// carver (blocks/) and never to the legacy standalone mirror (cas/). A remote
-	// that does not implement RemoteBlockStore leaves carve disabled and the
-	// legacy per-hash mirror in effect (back-compat).
-	if remoteStore != nil {
-		if rbs, ok := remoteStore.(remote.RemoteBlockStore); ok {
-			syncer.SetRemoteBlockStore(rbs)
-		}
-	}
-
-	cleanup := func() {
-		_ = syncer.Close()
-		_ = localStore.Close()
-		if remoteConfigID != "" {
-			s.releaseRemoteStore(remoteConfigID)
-		}
-	}
-
-	// Wire the metadata coordinator so the engine can invoke RefCount
-	// mutations + FileAttr.Blocks persistence without importing
-	// pkg/metadata on its hot paths. The fileChunkStore on the engine
-	// seam is the per-share metadata store cast to EngineFileChunkStore;
-	// the coordinator wraps the same store as a metadata.Store
-	// for the typed operations.
-	var coordinator engine.MetadataCoordinator
-	if metadataStore, ok := fileChunkStore.(metadata.Store); ok {
-		coordinator = newMetadataCoordinator(metadataStore)
-	}
-
-	engineCfg := engine.BlockStoreConfig{
-		Local:          localStore,
-		Remote:         engineRemote,
-		Syncer:         syncer,
-		FileChunkStore: fileChunkStore,
-		Coordinator:    coordinator,
-	}
-	// Thread the SyncedHashStore from the same per-share metadata
-	// backend the coordinator wraps so the engine's mirror-loop Flush
-	// can MarkSynced after each successful remote.Put. The interface
-	// check is the standard runtime-type narrowing used elsewhere in
-	// this factory (RollupStore, MetadataStore).
-	if shs, ok := fileChunkStore.(metadata.SyncedHashStore); ok {
-		engineCfg.SyncedHashStore = shs
-	}
-	if effectiveDefaults != nil {
-		engineCfg.ReadBufferBytes = effectiveDefaults.ReadBufferBytes
-	}
-	if syncerDefaults != nil {
-		engineCfg.PrefetchWorkers = syncerDefaults.PrefetchWorkers
-	}
-
-	bs, err := engine.New(engineCfg)
-	if err != nil {
-		cleanup()
-		return fmt.Errorf("failed to create BlockStore: %w", err)
-	}
-
-	// Apply the per-share durability tier (#1758), composed from the local store
-	// config. The "durability" enum (local|writeback|remote) selects the tier;
-	// when absent, the raw "require_durable_commit" (#1274) and "writeback"
-	// (#1757) bools are honored for backward compatibility. require_durable_commit
-	// is set before Start so it governs the very first commit; the metadata
-	// writeback flag is stashed and applied to the metadata service in AddShare
-	// (which holds the registrar), after RegisterStoreForShare.
-	if localStoreCfg, cfgErr := localCfg.GetConfig(); cfgErr == nil {
-		writeback, requireDurableCommit := resolveDurabilityTier(localStoreCfg, config.Name)
-		bs.SetRequireDurableCommit(requireDurableCommit)
-		share.writeback = writeback
-		// Durable tiers (local-durable and remote) verify warm reads per-record so
-		// on-disk corruption is caught and healed/failed-closed instead of returning
-		// silently-wrong bytes; the writeback tier keeps the raw fast read.
-		if vr, ok := localStore.(interface{ SetVerifyReads(bool) }); ok {
-			vr.SetVerifyReads(!writeback)
-		}
-	} else {
-		logger.Warn("failed to read local block store config for durability tier; defaulting to local",
-			"share", config.Name, "error", cfgErr)
-	}
-
-	if err := bs.Start(ctx); err != nil {
-		cleanup()
-		return fmt.Errorf("failed to start BlockStore: %w", err)
-	}
-
-	if err := seedColdIfNeeded(ctx, bs, localStore, fileChunkStore, remoteConfigured, config.Name); err != nil {
-		cleanup()
-		return err
-	}
-
-	// A local-only share (no remote to re-fetch from) that carried a complete
-	// pre-journal layout re-ingests its bytes from the surviving append logs in
-	// the BACKGROUND — AddShare must not block on O(total-bytes) work. Reads that
-	// arrive before a payload is drained fault it in per-payload (never zero-
-	// fill). When every payload is drained the archived legacy dirs are deleted;
-	// a crash before then leaves them on disk and the next open resumes. The
-	// final rollup is done here (outside any metadata txn — the coordinator is
-	// non-reentrant), never from inside a read.
-	if m, ok := localStore.(legacyLocalOnlyMigrator); ok && m.MigratedFromLegacyLocalOnly() {
-		shareName := config.Name
-		go func() {
-			// Detached from the AddShare context, which may be cancelled once the
-			// call returns; the store's own close gate stops the drain on shutdown.
-			bgCtx := context.Background()
-			pending := m.LegacyPendingPayloads()
-			started := time.Now()
-			lastLog := started
-			logger.Info("legacy local-only migration: re-ingesting archived append logs",
-				"share", shareName, "payloads_total", len(pending))
-			for i, payloadID := range pending {
-				if err := m.MaterializeLegacyPayload(payloadID); err != nil {
-					logger.Error("legacy local-only migration: materialize failed; leaving archive for retry on next start",
-						"share", shareName, "payload", payloadID, "error", err)
-					return
-				}
-				if time.Since(lastLog) >= migrationProgressInterval {
-					lastLog = time.Now()
-					logger.Info("legacy local-only migration: re-ingesting archived append logs",
-						"share", shareName, "payloads_done", i+1, "payloads_total", len(pending),
-						"elapsed", time.Since(started).Round(time.Second))
-				}
-			}
-			if err := bs.DrainRollups(bgCtx); err != nil {
-				logger.Error("legacy local-only migration: rollup drain failed; leaving archive for retry on next start",
-					"share", shareName, "error", err)
-				return
-			}
-			if err := m.FinishLegacyMigration(); err != nil {
-				logger.Error("legacy local-only migration: cleanup of archived legacy dirs failed",
-					"share", shareName, "error", err)
-				return
-			}
-			logger.Info("migrated pre-journal local-only layout: re-ingested bytes from append logs and removed the archive",
-				"share", shareName)
-		}()
-	}
-
-	// Thread the inline metrics recorder into the new store's eviction/
-	// backpressure path. nil when the runtime has not yet installed a handle
-	// (startup share-loading precedes metrics.New); SetMetrics back-fills
-	// those shares once it arrives. Read under mu — SetMetrics may run
-	// concurrently on another goroutine.
-	s.mu.RLock()
-	rec := s.metricsRec
-	s.mu.RUnlock()
-	if rec != nil {
-		bs.SetMetrics(rec)
-	}
-
-	// Safe without lock: share is not yet in the registry.
-	share.BlockStore = bs
-	share.remoteConfigID = remoteConfigID
-	// Compute the persistent gc-state directory for this share. Only fs-backed
-	// local stores produce a non-empty path; in-memory backends skip
-	// last-run.json persistence entirely (engine.PersistLastRunSummary is a
-	// no-op on empty rootDir).
-	share.gcStateRoot = deriveGCStateRoot(localCfg, config.Name)
-	// per-share local data dir for the migration journal.
-	// Same source-of-truth + emptiness semantics as gcStateRoot — memory
-	// backends produce "" so the status handler can short-circuit.
-	share.localStoreDir = deriveLocalStoreDir(localCfg, config.Name)
-
-	// A pinned share never evicts, so a bounded local tier can fill and then
-	// fail reads with ErrDiskFull once the working set exceeds it. Warn the
-	// operator at startup (no behavior change) so the misconfiguration is
-	// visible before it bites a client.
-	if config.RetentionPolicy == block.RetentionPin && remoteStore != nil && bs.LocalStats().MaxDisk > 0 {
-		logger.Warn("pinned share with a bounded local tier: reads will fail with ErrDiskFull once the working set exceeds the local tier — raise local_store_size or drop the pin",
-			"share", config.Name,
-			"local_store_size", bs.LocalStats().MaxDisk)
-	}
-
-	logger.Info("Per-share BlockStore initialized",
-		"share", config.Name,
-		"mode", modeLabel(remoteStore != nil),
-		"local_type", localCfg.Type,
-		"retention", config.RetentionPolicy,
-		"retention_ttl", config.RetentionTTL)
-
-	return nil
-}
-
-// RebindShareBlockStore hot-reloads a running share's per-share BlockStore so a
-// changed local/remote block-store binding takes effect WITHOUT a server
-// restart (see #1532). The syncer mode (local-only vs. remote) and the remote
-// target are fixed at BlockStore construction, so the only correct way to
-// change them on a live share is to tear the old store down and build a new one.
-//
-// The fs local backend is not safe to double-open, and for a remote-only change
-// the local directory is unchanged, so the old store is fully drained + closed
-// BEFORE the new one is constructed over the same dir. The share therefore has a
-// brief I/O gap during the swap: in-flight ops complete (Close drains them),
-// new ops briefly get ErrClosed and are retried by the NFS/SMB client. Binding
-// a remote also backfills pre-existing local blocks — the rebuilt syncer's
-// Start seeds the pending-upload set from disk.
-//
-// newConfig carries the new binding; oldConfig is identical except for the
-// block-store IDs and is used only to rebuild the previous store if the new one
-// fails to build, so a rebind failure never leaves the share storeless.
-func (s *Service) RebindShareBlockStore(
-	ctx context.Context,
-	newConfig *ShareConfig,
-	oldConfig *ShareConfig,
-	storeProvider MetadataStoreProvider,
-	blockStoreProvider BlockStoreConfigProvider,
-	localStoreDefaults *LocalStoreDefaults,
-	syncerDefaults *SyncerDefaults,
-) error {
-	name := newConfig.Name
-	if newConfig.LocalBlockStoreID == "" {
-		return fmt.Errorf("cannot rebind share %q: no local block store configured", name)
-	}
-
-	// Serialize rebinds: overlapping teardown/rebuild over the same local dir is
-	// unsafe.
-	s.rebindMu.Lock()
-	defer s.rebindMu.Unlock()
-
-	s.mu.RLock()
-	share, ok := s.registry[name]
-	s.mu.RUnlock()
-	if !ok {
-		return fmt.Errorf("cannot rebind share %q: not found in registry", name)
-	}
-
-	// Resolve the metadata store (fileChunkStore); unchanged by a binding change.
-	fileChunkStore, err := storeProvider.GetMetadataStore(newConfig.MetadataStore)
-	if err != nil {
-		return fmt.Errorf("failed to resolve metadata store for share %q: %w", name, err)
-	}
-
-	// Pre-validate the new binding resolves BEFORE tearing down the live store,
-	// so a bad binding fails fast without disrupting the running share.
-	if _, err := resolveBlockStoreConfig(ctx, blockStoreProvider, newConfig.LocalBlockStoreID, models.BlockStoreKindLocal); err != nil {
-		return fmt.Errorf("failed to resolve new local block store %q: %w", newConfig.LocalBlockStoreID, err)
-	}
-	if newConfig.RemoteBlockStoreID != "" {
-		if _, err := resolveBlockStoreConfig(ctx, blockStoreProvider, newConfig.RemoteBlockStoreID, models.BlockStoreKindRemote); err != nil {
-			return fmt.Errorf("failed to resolve new remote block store %q: %w", newConfig.RemoteBlockStoreID, err)
-		}
-	}
-
-	s.mu.RLock()
-	oldBS := share.BlockStore
-	oldRemoteConfigID := share.remoteConfigID
-	s.mu.RUnlock()
-
-	// Cancel any in-flight warm job for this share so it cannot keep fetching
-	// into the block store that is about to be drained and closed.
-	s.warmJobs.cancelForShare(name)
-
-	// Flush pending uploads to the OLD remote before teardown so switching or
-	// detaching a remote does not strand unmirrored blocks. Best-effort: a drain
-	// error must not block the rebind.
-	if oldBS != nil {
-		if err := oldBS.DrainAllUploads(ctx); err != nil {
-			logger.Warn("rebind: failed to drain uploads before teardown; continuing",
-				"share", name, "error", err)
-		}
-		if err := oldBS.Close(); err != nil {
-			logger.Warn("rebind: error closing previous block store; continuing",
-				"share", name, "error", err)
-		}
-	}
-
-	// Build the new store over the same (now-closed) local dir.
-	rebuilt := &Share{Name: name}
-	if buildErr := s.createBlockStoreForShare(ctx, rebuilt, newConfig, blockStoreProvider, fileChunkStore, localStoreDefaults, syncerDefaults); buildErr != nil {
-		// Recovery: rebuild the previous binding so the share is not left
-		// storeless. oldConfig differs from newConfig only in the block-store IDs.
-		logger.Error("rebind: failed to build new block store; restoring previous binding",
-			"share", name, "error", buildErr)
-		recovered := &Share{Name: name}
-		if recErr := s.createBlockStoreForShare(ctx, recovered, oldConfig, blockStoreProvider, fileChunkStore, localStoreDefaults, syncerDefaults); recErr != nil {
-			// Both failed: the share keeps its now-closed store (ops return
-			// ErrClosed, not a nil-deref panic) and needs a restart. Release the
-			// original remote ref since nothing holds it anymore.
-			if oldRemoteConfigID != "" {
-				s.releaseRemoteStore(oldRemoteConfigID)
-			}
-			return fmt.Errorf("rebind failed for share %q and previous binding could not be restored (%v); share needs a restart: %w", name, recErr, buildErr)
-		}
-		// Previous binding restored. Swap it in and drop the original remote ref
-		// (the recovery rebuild acquired its own) — but only if the share is
-		// still registered; a concurrent RemoveShare would already have released
-		// oldRemoteConfigID, so releasing it again here underflows the ref-count.
-		s.mu.Lock()
-		if cur, ok := s.registry[name]; !ok || cur != share {
-			s.mu.Unlock()
-			if closeErr := recovered.BlockStore.Close(); closeErr != nil {
-				logger.Warn("rebind: failed to close recovered block store after concurrent share removal",
-					"share", name, "error", closeErr)
-			}
-			if recovered.remoteConfigID != "" {
-				s.releaseRemoteStore(recovered.remoteConfigID)
-			}
-			return fmt.Errorf("share %q was removed during rebind (new binding also failed: %v)", name, buildErr)
-		}
-		share.BlockStore = recovered.BlockStore
-		share.remoteConfigID = recovered.remoteConfigID
-		share.gcStateRoot = recovered.gcStateRoot
-		share.localStoreDir = recovered.localStoreDir
-		s.blockStoreCache.Store(name, share.BlockStore)
-		s.mu.Unlock()
-		if oldRemoteConfigID != "" {
-			s.releaseRemoteStore(oldRemoteConfigID)
-		}
-		return fmt.Errorf("failed to rebind block store for share %q (previous binding restored): %w", name, buildErr)
-	}
-
-	// Swap the new store into the registry, but only if the share is still
-	// registered under the same pointer. A concurrent RemoveShare (which does
-	// not take rebindMu) can delete it and release oldRemoteConfigID while we
-	// rebuild; swapping into the stale pointer and releasing the old ref again
-	// would double-decrement the shared remote ref-count and could close a
-	// remote store still used by other shares.
-	s.mu.Lock()
-	if cur, ok := s.registry[name]; !ok || cur != share {
-		s.mu.Unlock()
-		// Share removed during rebind. Tear down the store we just built and
-		// drop its own remote ref; leave the old ref to RemoveShare.
-		if closeErr := rebuilt.BlockStore.Close(); closeErr != nil {
-			logger.Warn("rebind: failed to close new block store after concurrent share removal",
-				"share", name, "error", closeErr)
-		}
-		if rebuilt.remoteConfigID != "" {
-			s.releaseRemoteStore(rebuilt.remoteConfigID)
-		}
-		return fmt.Errorf("share %q was removed during rebind", name)
-	}
-	share.BlockStore = rebuilt.BlockStore
-	share.remoteConfigID = rebuilt.remoteConfigID
-	share.gcStateRoot = rebuilt.gcStateRoot
-	share.localStoreDir = rebuilt.localStoreDir
-	s.blockStoreCache.Store(name, share.BlockStore)
-	s.mu.Unlock()
-
-	// Drop the previous remote ref now that the new store holds its own.
-	if oldRemoteConfigID != "" {
-		s.releaseRemoteStore(oldRemoteConfigID)
-	}
-
-	s.notifyShareChange()
-	logger.Info("Per-share BlockStore rebound live",
-		"share", name,
-		"local_block_store_id", newConfig.LocalBlockStoreID,
-		"remote_block_store_id", newConfig.RemoteBlockStoreID,
-		"mode", modeLabel(newConfig.RemoteBlockStoreID != ""))
-	return nil
-}
-
-// acquireRemoteStore returns a shared remote store, creating it if needed.
-// Uses double-checked locking to avoid holding s.mu during potentially slow
-// network/DB I/O (config resolution, S3 client initialization).
-// Returns the store, its config ID, and any error.
-func (s *Service) acquireRemoteStore(ctx context.Context, ref string, provider BlockStoreConfigProvider) (remote.RemoteStore, string, error) {
-	// Fast path: when the share already persists the canonical UUID (the common
-	// case) and the store is live, take it without a config-resolution DB read.
-	// Legacy name refs (#1312) miss here — the map is keyed by UUID — and fall
-	// through to full resolution below.
-	s.mu.Lock()
-	if sr, ok := s.remoteStores[ref]; ok {
-		sr.refCount++
-		s.mu.Unlock()
-		return sr.store, ref, nil
-	}
-	s.mu.Unlock()
-
-	// Resolve config (by UUID, or by name for #1312 legacy rows) so the
-	// ref-count map is always keyed by the canonical store UUID. Two shares
-	// referencing the same remote — one by UUID, one by legacy name — must
-	// share the single ref-counted store.
-	remoteCfg, err := resolveBlockStoreConfig(ctx, provider, ref, models.BlockStoreKindRemote)
-	if err != nil {
-		return nil, "", fmt.Errorf("failed to resolve remote block store config %q: %w", ref, err)
-	}
-	if remoteCfg.Kind != models.BlockStoreKindRemote {
-		return nil, "", fmt.Errorf("block store config %q has kind %q, expected %q", ref, remoteCfg.Kind, models.BlockStoreKindRemote)
-	}
-	configID := remoteCfg.ID
-
-	s.mu.Lock()
-	if sr, ok := s.remoteStores[configID]; ok {
-		sr.refCount++
-		s.mu.Unlock()
-		return sr.store, configID, nil
-	}
-	s.mu.Unlock()
-
-	newStore, err := CreateRemoteStoreFromConfig(ctx, remoteCfg.Type, remoteCfg)
-	if err != nil {
-		return nil, "", fmt.Errorf("failed to create remote store: %w", err)
-	}
-
-	// Decorator order matters: encryption sits BELOW compression on the
-	// data flow (caller → compression → encryption → inner). Compress
-	// plaintext first so the compressor sees redundancy; encrypted bytes
-	// are incompressible by design.
-	//
-	// Apply order in code is therefore encryption first (innermost),
-	// then compression (outermost).
-	encWrapped, err := maybeWrapEncryption(ctx, newStore, remoteCfg)
-	if err != nil {
-		_ = newStore.Close()
-		return nil, "", fmt.Errorf("failed to apply encryption policy: %w", err)
-	}
-	newStore = encWrapped
-
-	wrapped, err := maybeWrapCompression(newStore, remoteCfg)
-	if err != nil {
-		_ = newStore.Close()
-		return nil, "", fmt.Errorf("failed to apply compression policy: %w", err)
-	}
-	newStore = wrapped
-
-	// Double-check: another goroutine may have created the store concurrently.
-	s.mu.Lock()
-	if sr, ok := s.remoteStores[configID]; ok {
-		sr.refCount++
-		s.mu.Unlock()
-		// We lost the race; discard our now-redundant store. newStore is the
-		// fully-decorated (encryption/compression) stack, so a Close failure
-		// here could leak a key provider — surface it rather than swallow it.
-		if err := newStore.Close(); err != nil {
-			logger.Warn("acquireRemoteStore: failed to close duplicate remote store",
-				"config_id", configID, "error", err)
-		}
-		return sr.store, configID, nil
-	}
-
-	s.remoteStores[configID] = &sharedRemote{
-		store:    newStore,
-		refCount: 1,
-	}
-	s.mu.Unlock()
-
-	logger.Info("Created shared remote store", "config_id", configID, "type", remoteCfg.Type)
-	return newStore, configID, nil
-}
-
-// maybeWrapEncryption inspects the remote config's "encryption" key and,
-// when present, wraps inner with an encryption.EncryptedRemote. Returns
-// inner unchanged when the key is absent.
-//
-// Key-provider lifetime is bound to the decorator: NewRemote captures
-// the provider, and EncryptedRemote.Close calls provider.Close. The
-// outer releaseRemoteStore path therefore closes the provider as part
-// of the normal decorator teardown.
-func maybeWrapEncryption(ctx context.Context, inner remote.RemoteStore, cfg *models.BlockStoreConfig) (remote.RemoteStore, error) {
-	parsed, err := cfg.GetConfig()
-	if err != nil {
-		return nil, fmt.Errorf("parse block store config: %w", err)
-	}
-	raw, ok := parsed["encryption"]
-	if !ok {
-		return inner, nil
-	}
-	encoded, err := json.Marshal(raw)
-	if err != nil {
-		return nil, fmt.Errorf("marshal encryption sub-config: %w", err)
-	}
-	policy, err := encryption.ParsePolicy(encoded)
-	if err != nil {
-		return nil, err
-	}
-	provider, err := keyprovider.NewProvider(ctx, policy.Key)
-	if err != nil {
-		return nil, fmt.Errorf("create key provider: %w", err)
-	}
-	wrapped, err := encryption.NewRemote(inner, policy, provider)
-	if err != nil {
-		_ = provider.Close()
-		return nil, err
-	}
-	return wrapped, nil
-}
-
-// maybeWrapCompression inspects the remote config's "compression" key
-// and, when present, wraps inner with a compression.Decorator. Returns
-// inner unchanged when the key is absent.
-func maybeWrapCompression(inner remote.RemoteStore, cfg *models.BlockStoreConfig) (remote.RemoteStore, error) {
-	parsed, err := cfg.GetConfig()
-	if err != nil {
-		return nil, fmt.Errorf("parse block store config: %w", err)
-	}
-	raw, ok := parsed["compression"]
-	if !ok {
-		return inner, nil
-	}
-	encoded, err := json.Marshal(raw)
-	if err != nil {
-		return nil, fmt.Errorf("marshal compression sub-config: %w", err)
-	}
-	policy, err := compression.ParsePolicy(encoded)
-	if err != nil {
-		return nil, err
-	}
-	return compression.NewRemote(inner, policy)
-}
-
-// releaseRemoteStore decrements the reference count and closes the remote store if no longer used.
-// Close happens outside the lock to avoid blocking share operations during network I/O.
-func (s *Service) releaseRemoteStore(configID string) {
-	var storeToClose remote.RemoteStore
-
-	s.mu.Lock()
-	sr, ok := s.remoteStores[configID]
-	if !ok {
-		s.mu.Unlock()
-		return
-	}
-	sr.refCount--
-	if sr.refCount <= 0 {
-		storeToClose = sr.store
-		delete(s.remoteStores, configID)
-	}
-	s.mu.Unlock()
-
-	if storeToClose != nil {
-		_ = storeToClose.Close()
-		logger.Info("Closed shared remote store", "config_id", configID)
-	}
-}
-
-// RemoveShare removes a share from the registry and closes its BlockStore.
-// Does not close the underlying metadata store.
-//
-// Teardown is ordered best-effort (REVIEW M4): every step runs even if an
-// earlier one fails, and the per-step errors are aggregated into the returned
-// error so the registry / snapshot-dir / block-store / remote-ref state is
-// never left half-removed by an early return. The registry entry is dropped
-// first (under the lock) so the share disappears from routing immediately;
-// the remaining steps are pure resource teardown.
-//
-// bs.Close() is now drain-safe: it takes the engine's lifecycle
-// write lock, which blocks until all in-flight WriteAt/ReadAt/Flush ops on the
-// store have completed, so calling it here outside s.mu can no longer race a
-// client mid-transfer into a torn op or panic.
-func (s *Service) RemoveShare(name string) error {
-	s.mu.Lock()
-	share, exists := s.registry[name]
-	if !exists {
-		s.mu.Unlock()
-		return fmt.Errorf("share %q not found", name)
-	}
-	bs := share.BlockStore
-	remoteConfigID := share.remoteConfigID
-	localStoreDir := share.localStoreDir
-	delete(s.registry, name)
-	s.blockStoreCache.Delete(name)
-	s.mu.Unlock()
-
-	// Cancel any in-flight warm job for this share so it cannot keep fetching
-	// into a block store that is about to be closed.
-	s.warmJobs.cancelForShare(name)
-
-	var errs []error
-
-	// Cleanup per-share snapshot directories alongside registry removal.
-	// The DB row is the source of truth; orphaned files left behind on a
-	// removal error are operationally harmless, so we log + aggregate but
-	// continue with the rest of the teardown.
-	if localStoreDir != "" {
-		snapsDir := filepath.Join(localStoreDir, "snapshots")
-		if err := os.RemoveAll(snapsDir); err != nil {
-			logger.Warn("RemoveShare: failed to remove snapshots dir",
-				"share", name, "dir", snapsDir, "error", err)
-			errs = append(errs, fmt.Errorf("remove snapshots dir %q: %w", snapsDir, err))
-		}
-	}
-
-	if bs != nil {
-		if err := bs.Close(); err != nil {
-			logger.Warn("Failed to close BlockStore for share", "share", name, "error", err)
-			errs = append(errs, fmt.Errorf("close block store for share %q: %w", name, err))
-		}
-	}
-
-	// Always release the remote-store reference even if a prior step errored,
-	// otherwise a Close failure would leak the shared remote ref-count.
-	if remoteConfigID != "" {
-		s.releaseRemoteStore(remoteConfigID)
-	}
-
-	s.notifyShareChange()
-
-	return errors.Join(errs...)
-}
-
-// StopRollups stops and drains the rollup worker pool of every registered
-// share's block store. The runtime calls this during shutdown BEFORE it closes
-// the metadata stores (#1543): the rollup ticker persists FileChunk manifests
-// and rollup offsets through the metadata store, so it must be fenced first or
-// an in-flight rollup races the DB close ("sql: database is closed") and can
-// drop a local chunk that was never mirrored.
-//
-// The ctx bounds the TOTAL drain time (an overall deadline): each store is
-// given the time remaining until ctx's deadline as its grace, so shutdown stays
-// bounded regardless of share count. Once the budget is spent the worker-pool
-// fence still runs (that is the load-bearing part — it stops the ticker); only
-// the best-effort drain is skipped, and those intervals resume on restart.
-//
-// Best-effort — a per-share drain error is logged, not propagated, so one share
-// cannot block the rest of shutdown. Drains run outside the registry lock (a
-// drain can block up to its grace window). The block stores stay OPEN; their
-// full teardown still happens in RemoveShare.
-func (s *Service) StopRollups(ctx context.Context) {
-	type namedStore struct {
-		name string
-		bs   *engine.Store
-	}
-	s.mu.RLock()
-	stores := make([]namedStore, 0, len(s.registry))
-	for name, share := range s.registry {
-		if share.BlockStore != nil {
-			stores = append(stores, namedStore{name: name, bs: share.BlockStore})
-		}
-	}
-	s.mu.RUnlock()
-
-	for _, ns := range stores {
-		// grace = time left until the shared deadline (overall bound). No
-		// deadline → 0, which defers to the store's default. Budget already
-		// spent → a 1ms floor so we still fence the pool without reviving the
-		// 30s default that GracefulStopRollup applies to grace <= 0.
-		grace := time.Duration(0)
-		if dl, ok := ctx.Deadline(); ok {
-			if grace = time.Until(dl); grace <= 0 {
-				grace = time.Millisecond
-			}
-		}
-		if err := ns.bs.StopRollup(grace); err != nil {
-			logger.Warn("Failed to stop rollup for share; remaining rollups resume on restart",
-				"share", ns.name, "error", err)
-		}
-	}
-}
-
-func (s *Service) UpdateShare(name string, readOnly *bool, defaultPermission *string, retentionPolicy *block.RetentionPolicy, retentionTTL *time.Duration) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	share, exists := s.registry[name]
-	if !exists {
-		return fmt.Errorf("share %q not found", name)
-	}
-
-	if readOnly != nil {
-		share.ReadOnly = *readOnly
-	}
-	if defaultPermission != nil {
-		share.DefaultPermission = *defaultPermission
-	}
-	if retentionPolicy != nil {
-		share.RetentionPolicy = *retentionPolicy
-	}
-	if retentionTTL != nil {
-		share.RetentionTTL = *retentionTTL
-	}
-
-	// Only the pin/non-pin distinction reaches the block store: the local tier
-	// evicts whole fully-synced segments approx-LRU and has no ttl or lru knob to
-	// receive, so the rest of the policy is metadata the API reports back.
-	if (retentionPolicy != nil || retentionTTL != nil) && share.BlockStore != nil {
-		// Pin mode disables eviction; switching away from pin re-enables it
-		// (unless the share is local-only, in which case eviction stays disabled).
-		if share.RetentionPolicy == block.RetentionPin {
-			share.BlockStore.SetEvictionEnabled(false)
-		} else if share.BlockStore.HasRemoteStore() {
-			share.BlockStore.SetEvictionEnabled(true)
-		}
-	}
-
-	return nil
-}
-
-// SetShareSquash updates the live in-memory squash policy (and optionally the
-// anonymous UID/GID) for a share, so an NFS squash-config change applies to
-// active clients without an adapter restart. The on-disk config is persisted
-// separately by the API handler; this only refreshes the runtime Share that
-// ResolveSharePermission / ApplyIdentityMapping read from. anonUID/anonGID are
-// pointers so an unspecified field is left unchanged. Mirrors UpdateShare's
-// lock+mutate pattern; GetShare hands callers a snapshot copy, so mutating the
-// registry entry under the lock is race-free.
-func (s *Service) SetShareSquash(name string, squash models.SquashMode, anonUID, anonGID *uint32) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	share, exists := s.registry[name]
-	if !exists {
-		return fmt.Errorf("share %q not found", name)
-	}
-
-	share.Squash = squash
-	if anonUID != nil {
-		share.AnonymousUID = *anonUID
-	}
-	if anonGID != nil {
-		share.AnonymousGID = *anonGID
-	}
-
-	return nil
-}
-
-// TrashSettings is a per-share recycle-bin policy snapshot, returned by value
-// under the service lock so callers never read a mutating shared pointer.
-type TrashSettings struct {
-	Enabled         bool
-	RetentionDays   int
-	RestrictToAdmin bool
-	MaxBytes        int64
-	ExcludePatterns []string
-}
-
-// TrashSettingsForShare returns the recycle-bin policy for a share, read under
-// the service lock. Returns ok=false if the share is unknown. Never hands out
-// the *Share; the ExcludePatterns slice is copied so callers cannot observe a
-// concurrent SetShareTrashConfig mutation.
-//
-// The recycle decision is read per-delete from many protocol goroutines while
-// config can be updated live, so this accessor returns a VALUE under the lock
-// (refs #190, #936 race lesson).
-func (s *Service) TrashSettingsForShare(name string) (TrashSettings, bool) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	share, exists := s.registry[name]
-	if !exists {
-		return TrashSettings{}, false
-	}
-	return TrashSettings{
-		Enabled:         share.TrashEnabled,
-		RetentionDays:   share.TrashRetentionDays,
-		RestrictToAdmin: share.TrashRestrictToAdmin,
-		MaxBytes:        share.TrashMaxBytes,
-		ExcludePatterns: append([]string(nil), share.TrashExcludePatterns...),
-	}, true
-}
-
-// EnabledTrashShares returns the names of all shares with trash enabled, read
-// under the service lock. Used by the reaper loop to enumerate bins to sweep.
-func (s *Service) EnabledTrashShares() []string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	var out []string
-	for name, share := range s.registry {
-		if share.TrashEnabled {
-			out = append(out, name)
-		}
-	}
-	return out
-}
-
-// SetShareTrashConfig updates a live share's recycle-bin settings under the
-// write lock and persists them to the DB. Pairs safely with the RLock
-// TrashSettingsForShare accessor: the exact fields the reader copies are the
-// ones this mutates while holding the write lock.
-//
-// Returns ErrShareNotFound if the share name is unknown in the runtime
-// registry. The optional store argument persists the change; pass nil to
-// update runtime state only (used by tests that bypass the DB).
-func (s *Service) SetShareTrashConfig(store ShareStore, name string, cfg TrashSettings) error {
-	excludes := append([]string(nil), cfg.ExcludePatterns...)
-
-	s.mu.Lock()
-	share, exists := s.registry[name]
-	if !exists {
-		s.mu.Unlock()
-		return fmt.Errorf("%w: %q", ErrShareNotFound, name)
-	}
-	share.TrashEnabled = cfg.Enabled
-	share.TrashRetentionDays = cfg.RetentionDays
-	share.TrashRestrictToAdmin = cfg.RestrictToAdmin
-	share.TrashMaxBytes = cfg.MaxBytes
-	share.TrashExcludePatterns = excludes
-	s.mu.Unlock()
-
-	if store == nil {
-		return nil
-	}
-
-	dbShare, err := store.GetShare(context.Background(), name)
-	if err != nil {
-		return fmt.Errorf("load share %q: %w", name, err)
-	}
-	dbShare.TrashEnabled = cfg.Enabled
-	dbShare.TrashRetentionDays = cfg.RetentionDays
-	dbShare.TrashRestrictToAdmin = cfg.RestrictToAdmin
-	dbShare.TrashMaxBytes = cfg.MaxBytes
-	dbShare.SetTrashExcludePatterns(excludes)
-	if err := store.UpdateShare(context.Background(), dbShare); err != nil {
-		return fmt.Errorf("persist trash config for share %q: %w", name, err)
-	}
-	return nil
-}
-
-// SetShareNetgroup updates the live netgroup association for a share's NFS
-// export. An empty netgroupName clears the association (allow-all). The change
-// takes effect immediately: CheckNetgroupAccess reads NetgroupName from this
-// runtime registry, so subsequent NFS operations honour the new allowlist
-// without an adapter restart. Returns an error if the share is unknown.
-func (s *Service) SetShareNetgroup(name, netgroupName string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	share, exists := s.registry[name]
-	if !exists {
-		return fmt.Errorf("%w: %q", ErrShareNotFound, name)
-	}
-	share.NetgroupName = netgroupName
-	return nil
-}
-
-// GetShareNetgroupName returns the live netgroup association for a share,
-// read under s.mu. Callers (e.g. CheckNetgroupAccess) must use this rather
-// than reading NetgroupName off a *Share returned by GetShare: GetShare hands
-// back the shared registry pointer with the lock already dropped, so reading
-// the field there races with SetShareNetgroup's write under s.mu. The bool is
-// false when the share is unknown.
-func (s *Service) GetShareNetgroupName(name string) (string, bool) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	share, exists := s.registry[name]
-	if !exists {
-		return "", false
-	}
-	return share.NetgroupName, true
-}
-
-// DisableShare sets enabled=false on the share's DB row and runtime Share
-// struct, then invokes notifyShareChange so adapters drop active sessions.
-// DB-first-then-runtime ordering is crash-consistent: if the process dies
-// between the two, the next boot reconciles runtime from DB.
-//
-// Idempotent: re-calling on an already-disabled share returns
-// ErrShareAlreadyDisabled without writing to DB or disturbing adapters.
-//
-// Returns ErrShareNotFound if the share name is unknown at either layer.
-// Timeout bound is whatever the caller provides via ctx.
-//
-// Requires `"enabled"` in GORMStore.UpdateShare's update whitelist —
-// otherwise the store.UpdateShare call silently drops the flag.
-func (s *Service) DisableShare(ctx context.Context, store ShareStore, name string) error {
-	// Runtime registry must know the share before we touch the DB — prevents
-	// a DB-disabled/runtime-absent inconsistency when the startup load missed
-	// a share (partial boot) or the caller passed a stale name.
-	s.mu.RLock()
-	_, exists := s.registry[name]
-	s.mu.RUnlock()
-	if !exists {
-		return fmt.Errorf("%w: runtime registry: %q", ErrShareNotFound, name)
-	}
-
-	dbShare, err := store.GetShare(ctx, name)
-	if err != nil {
-		return fmt.Errorf("load share %q: %w", name, err)
-	}
-	if !dbShare.Enabled {
-		return ErrShareAlreadyDisabled
-	}
-	dbShare.Enabled = false
-	if err := store.UpdateShare(ctx, dbShare); err != nil {
-		return fmt.Errorf("persist disabled state for share %q: %w", name, err)
-	}
-
-	s.mu.Lock()
-	share, stillExists := s.registry[name]
-	if !stillExists {
-		s.mu.Unlock()
-		return fmt.Errorf("%w: runtime registry: %q", ErrShareNotFound, name)
-	}
-	share.Enabled = false
-	s.mu.Unlock()
-
-	s.notifyShareChange()
-	return nil
-}
-
-// EnableShare inverts DisableShare. Idempotent: re-calling on an
-// already-enabled share is a no-op (returns nil, no DB write).
-func (s *Service) EnableShare(ctx context.Context, store ShareStore, name string) error {
-	// Registry-first check: same rationale as DisableShare — avoid a DB row
-	// that moves while the runtime has no matching entry.
-	s.mu.RLock()
-	_, exists := s.registry[name]
-	s.mu.RUnlock()
-	if !exists {
-		return fmt.Errorf("%w: runtime registry: %q", ErrShareNotFound, name)
-	}
-
-	dbShare, err := store.GetShare(ctx, name)
-	if err != nil {
-		return fmt.Errorf("load share %q: %w", name, err)
-	}
-	if dbShare.Enabled {
-		return nil
-	}
-	dbShare.Enabled = true
-	if err := store.UpdateShare(ctx, dbShare); err != nil {
-		return fmt.Errorf("persist enabled state for share %q: %w", name, err)
-	}
-
-	s.mu.Lock()
-	share, stillExists := s.registry[name]
-	if !stillExists {
-		s.mu.Unlock()
-		return fmt.Errorf("%w: runtime registry: %q", ErrShareNotFound, name)
-	}
-	share.Enabled = true
-	s.mu.Unlock()
-
-	s.notifyShareChange()
-	return nil
-}
-
-// IsShareEnabled returns the runtime Enabled flag for the named share.
-// Mirror of GetShare read-path discipline (RLock + registry lookup).
-func (s *Service) IsShareEnabled(name string) (bool, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	share, exists := s.registry[name]
-	if !exists {
-		return false, fmt.Errorf("%w: %q", ErrShareNotFound, name)
-	}
-	return share.Enabled, nil
-}
-
-// ListEnabledSharesForStore returns the names of all runtime shares that
-// (a) have Enabled=true AND (b) reference metadataStoreName as their
-// metadata store.
-func (s *Service) ListEnabledSharesForStore(metadataStoreName string) []string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	var out []string
-	for name, share := range s.registry {
-		if share.Enabled && share.MetadataStore == metadataStoreName {
-			out = append(out, name)
-		}
-	}
-	return out
-}
-
 func (s *Service) GetShare(name string) (*Share, error) {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
@@ -2130,133 +409,6 @@ func (s *Service) LocalStoreDir(name string) (string, error) {
 	}
 	return share.localStoreDir, nil
 }
-
-// RegisterShareForTesting inserts a minimal enabled Share into the registry
-// without the full AddShare composition (no block/remote stores, no DB row).
-// Test-only — used by handler unit tests that register a metadata store via
-// MetadataService.RegisterStoreForShare and need the share to also resolve in
-// identity mapping and permission lookups (the production invariant: every
-// share with a metadata store is a registered share). If the share already
-// exists it is left untouched.
-func (s *Service) RegisterShareForTesting(name string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if _, exists := s.registry[name]; exists {
-		return
-	}
-	s.registry[name] = &Share{
-		Name:              name,
-		Enabled:           true,
-		DefaultPermission: "read-write",
-		Squash:            models.SquashNone,
-		// Mirror the AddShare default (allowAuthSys defaults true when unset):
-		// a bare test share must permit AUTH_SYS, otherwise the v4 export
-		// auth-flavor policy denies every AUTH_UNIX op.
-		AllowAuthSys: true,
-	}
-}
-
-// SetLocalStoreDirForTesting overrides the per-share localStoreDir field.
-// Test-only — used by handler unit tests that bypass the full
-// AddShare composition path (which requires a DB-backed
-// BlockStoreConfig). Returns ErrShareNotFound if the share is not
-// registered.
-func (s *Service) SetLocalStoreDirForTesting(name, dir string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	share, ok := s.registry[name]
-	if !ok {
-		return fmt.Errorf("%w: %q", ErrShareNotFound, name)
-	}
-	share.localStoreDir = dir
-	return nil
-}
-
-// SetBlockStoreForTesting overrides the per-share BlockStore field under the
-// service lock. Test-only — used by tests that compose a minimal BlockStore
-// out of band and must publish it into the registry. Production code sets
-// BlockStore only during AddShare. Returns ErrShareNotFound if the share is
-// not registered.
-//
-// Use this instead of mutating a *Share returned by GetShare: GetShare now
-// hands back a snapshot copy, so registry state must be changed through a
-// locked setter.
-func (s *Service) SetBlockStoreForTesting(name string, bs *engine.Store) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	share, ok := s.registry[name]
-	if !ok {
-		return fmt.Errorf("%w: %q", ErrShareNotFound, name)
-	}
-	share.BlockStore = bs
-	return nil
-}
-
-// SetEnabledForTesting overrides the per-share Enabled flag under the service
-// lock. Test-only — production code flips Enabled via DisableShare/EnableShare
-// (which also persist the DB row). Use this instead of mutating a *Share
-// returned by GetShare, which is now a snapshot copy. Returns ErrShareNotFound
-// if the share is not registered.
-func (s *Service) SetEnabledForTesting(name string, enabled bool) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	share, ok := s.registry[name]
-	if !ok {
-		return fmt.Errorf("%w: %q", ErrShareNotFound, name)
-	}
-	share.Enabled = enabled
-	return nil
-}
-
-// SetSharePolicyForTesting overrides a share's DefaultPermission and Squash mode
-// in the registry under lock. Test-only — lets NFS auth tests exercise the
-// export-squash permission policy without a full AddShare flow.
-func (s *Service) SetSharePolicyForTesting(name, defaultPermission string, squash models.SquashMode) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	share, ok := s.registry[name]
-	if !ok {
-		return fmt.Errorf("%w: %q", ErrShareNotFound, name)
-	}
-	share.DefaultPermission = defaultPermission
-	share.Squash = squash
-	return nil
-}
-
-// SetExportAuthPolicyForTesting overrides a share's AllowAuthSys and
-// RequireKerberos export auth-flavor fields in the registry under lock.
-// Test-only — lets NFS auth tests exercise the per-share export auth-flavor
-// policy (NFS4ERR_WRONGSEC enforcement) without a full AddShare flow. Use this
-// instead of mutating a *Share returned by GetShare, which is now a snapshot
-// copy. Returns ErrShareNotFound if the share is not registered.
-func (s *Service) SetExportAuthPolicyForTesting(name string, allowAuthSys, requireKerberos bool) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	share, ok := s.registry[name]
-	if !ok {
-		return fmt.Errorf("%w: %q", ErrShareNotFound, name)
-	}
-	share.AllowAuthSys = allowAuthSys
-	share.RequireKerberos = requireKerberos
-	return nil
-}
-
-// SetMinKerberosLevelForTesting overrides a share's MinKerberosLevel (the GSS
-// protection floor: "", "krb5", "krb5i", "krb5p") in the registry under lock.
-// Test-only — lets NFS auth tests exercise the min_kerberos_level floor
-// (NFS4ERR_WRONGSEC / MOUNT refusal) without a full AddShare flow. Returns
-// ErrShareNotFound if the share is not registered.
-func (s *Service) SetMinKerberosLevelForTesting(name, minLevel string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	share, ok := s.registry[name]
-	if !ok {
-		return fmt.Errorf("%w: %q", ErrShareNotFound, name)
-	}
-	share.MinKerberosLevel = minLevel
-	return nil
-}
-
 func (s *Service) GetRootHandle(shareName string) (metadata.FileHandle, error) {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
@@ -2267,7 +419,6 @@ func (s *Service) GetRootHandle(shareName string) (metadata.FileHandle, error) {
 	}
 	return share.RootHandle, nil
 }
-
 func (s *Service) ListShares() []string {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
@@ -2301,7 +452,6 @@ func (s *Service) SetMetrics(rec local.MetricsRecorder) {
 		bs.SetMetrics(rec)
 	}
 }
-
 func (s *Service) ShareExists(name string) bool {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
@@ -2377,7 +527,6 @@ func (s *Service) InvalidateAuthCache() {
 		cb()
 	}
 }
-
 func (s *Service) GetShareNameForHandle(ctx context.Context, handle metadata.FileHandle) (string, error) {
 	shareName, _, err := metadata.DecodeFileHandle(handle)
 	if err != nil {
@@ -2394,23 +543,10 @@ func (s *Service) GetShareNameForHandle(ctx context.Context, handle metadata.Fil
 
 	return shareName, nil
 }
-
 func (s *Service) CountShares() int {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
 	return len(s.registry)
-}
-
-// GetBlockStoreForHandle decodes a file handle and resolves the per-share
-// BlockStore. The handle is decoded once for its share name, then the store is
-// read from the lock-free blockStoreCache — the read/write hot path never
-// touches s.mu.
-func (s *Service) GetBlockStoreForHandle(_ context.Context, handle metadata.FileHandle) (*engine.Store, error) {
-	shareName, _, err := metadata.DecodeFileHandle(handle)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode file handle: %w", err)
-	}
-	return s.GetBlockStoreForShare(shareName)
 }
 
 // XattrStreamReader returns a metadata.StreamContentReader that materialises a
@@ -2441,931 +577,5 @@ func (s *Service) XattrStreamReader() metadata.StreamContentReader {
 			return nil, rerr
 		}
 		return buf[:n], nil
-	}
-}
-
-// GetBlockStoreForShare returns the BlockStore for a named share. The common
-// case is served from the lock-free blockStoreCache; a miss falls through to
-// the locked registry, which produces the exact not-found / no-store errors.
-// Because the cache is written under s.mu alongside the registry, a miss after
-// RemoveShare always resolves to a not-found error — never a stale or closed
-// store.
-func (s *Service) GetBlockStoreForShare(name string) (*engine.Store, error) {
-	if v, ok := s.blockStoreCache.Load(name); ok {
-		if bs, ok := v.(*engine.Store); ok {
-			return bs, nil
-		}
-	}
-
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	share, exists := s.registry[name]
-	if !exists {
-		return nil, fmt.Errorf("share %q not found", name)
-	}
-	if share.BlockStore == nil {
-		return nil, fmt.Errorf("share %q has no block store configured", name)
-	}
-	return share.BlockStore, nil
-}
-
-// RemoteStoreEntry describes a distinct remote block store that is referenced
-// by one or more shares. Surface used by production block-GC enumeration
-// (Runtime.RunBlockGC): we want each underlying remote store scanned exactly
-// once per run, not once per share.
-type RemoteStoreEntry struct {
-	// Store is the underlying remote store (NOT the nonClosingRemote wrapper).
-	Store remote.RemoteStore
-	// ConfigID is the remote block-store config UUID this entry represents.
-	// Empty string indicates a test-only binding (SetShareRemoteForTest).
-	ConfigID string
-	// Shares are the registered share names that reference this remote.
-	Shares []string
-}
-
-// DistinctRemoteStores returns every distinct underlying remote.RemoteStore
-// referenced by at least one registered share. Shares that reference the same
-// remote-store config (ref-counted via remoteStores) are grouped into a
-// single entry — deduped by ConfigID, NOT by the per-share nonClosingRemote
-// wrapper pointer. Local-only shares (no remote) contribute nothing.
-//
-// Returned entries have a non-nil Store and a non-empty Shares slice. Order
-// is unspecified (map iteration).
-func (s *Service) DistinctRemoteStores() []RemoteStoreEntry {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	// Bucket share names by the configID they reference. configID == "" means
-	// "local-only share" — skipped entirely.
-	sharesByConfigID := make(map[string][]string, len(s.remoteStores))
-	for name, sh := range s.registry {
-		if sh.remoteConfigID == "" {
-			continue
-		}
-		sharesByConfigID[sh.remoteConfigID] = append(sharesByConfigID[sh.remoteConfigID], name)
-	}
-
-	out := make([]RemoteStoreEntry, 0, len(sharesByConfigID))
-	for cid, shareNames := range sharesByConfigID {
-		sr, ok := s.remoteStores[cid]
-		if !ok || sr == nil || sr.store == nil {
-			// Orphaned configID → skip. DistinctRemoteStores is a read-only
-			// surface; we don't try to self-heal bookkeeping here.
-			continue
-		}
-		out = append(out, RemoteStoreEntry{
-			Store:    sr.store,
-			ConfigID: cid,
-			Shares:   shareNames,
-		})
-	}
-	return out
-}
-
-// LocalStoreEntry describes one share's local block store for a per-share local
-// GC pass. Unlike remote stores, local stores are never shared across shares
-// (architecture invariant #4), so there is exactly one entry per share and no
-// deduplication is needed.
-type LocalStoreEntry struct {
-	// ShareName is the single share that owns this local store.
-	ShareName string
-	// Store is the share's local block store. The local GC pass uses only its
-	// Walk + Delete methods.
-	Store block.Store
-	// GCStateRoot is the persistent gc-state directory for this share. Empty
-	// for in-memory backends (no on-disk chunks worth sweeping).
-	GCStateRoot string
-}
-
-// ShareLocalStores returns one LocalStoreEntry per registered share that has a
-// block store with a non-nil local tier. The local GC pass sweeps each share's
-// own local store against that share's live set (per-share isolation).
-func (s *Service) ShareLocalStores() []LocalStoreEntry {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	out := make([]LocalStoreEntry, 0, len(s.registry))
-	for name, sh := range s.registry {
-		if sh.BlockStore == nil {
-			continue
-		}
-		local := sh.BlockStore.LocalStore()
-		if local == nil {
-			continue
-		}
-		out = append(out, LocalStoreEntry{
-			ShareName:   name,
-			Store:       local,
-			GCStateRoot: sh.gcStateRoot,
-		})
-	}
-	return out
-}
-
-// SetShareRemoteForTest installs a remote.RemoteStore for the named share
-// and registers it under a synthetic configID derived from the store's
-// pointer identity. Two calls with the same remote store for different
-// shares share one configID — matching production ref-counting behavior
-// — so DistinctRemoteStores() dedupes correctly.
-//
-// Test-only: panics if the share does not exist. Intended for runtime-
-// package tests that need to exercise RunBlockGC's enumeration without
-// standing up a full engine.Store. Not safe for production callers.
-func (s *Service) SetShareRemoteForTest(shareName string, rs remote.RemoteStore) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	share, ok := s.registry[shareName]
-	if !ok {
-		panic(fmt.Sprintf("SetShareRemoteForTest: share %q not registered", shareName))
-	}
-	// Derive a stable configID from the remote store pointer so calls that
-	// pass the same rs for different shares land in the same sharedRemote
-	// bucket (mirroring production ref-count semantics).
-	cid := fmt.Sprintf("test-remote-%p", rs)
-	if existing, ok := s.remoteStores[cid]; ok {
-		existing.refCount++
-	} else {
-		s.remoteStores[cid] = &sharedRemote{
-			store:    rs,
-			refCount: 1,
-		}
-	}
-	share.remoteConfigID = cid
-}
-
-// DrainAllBlockStores drains all pending uploads across all per-share BlockStores.
-func (s *Service) DrainAllBlockStores(ctx context.Context) error {
-	s.mu.RLock()
-	blockStores := make([]*engine.Store, 0, len(s.registry))
-	for _, share := range s.registry {
-		if share.BlockStore != nil {
-			blockStores = append(blockStores, share.BlockStore)
-		}
-	}
-	s.mu.RUnlock()
-
-	var errs []error
-	for _, bs := range blockStores {
-		if err := bs.DrainAllUploads(ctx); err != nil {
-			errs = append(errs, err)
-		}
-	}
-	return errors.Join(errs...)
-}
-
-// UploadProgress returns a monotonic count of mirror attempts that have
-// concluded (completed + failed) summed across every share's block store.
-//
-// It is a liveness signal for the drain-uploads idle watchdog: while a drain
-// is genuinely making progress this value keeps rising; a value that stays
-// flat across an idle window means the drain has stalled (e.g. the remote went
-// unreachable). Counting failed attempts as well as completed ones means a
-// retrying upload still registers as activity, so transient errors don't trip
-// the watchdog.
-func (s *Service) UploadProgress() int64 {
-	s.mu.RLock()
-	blockStores := make([]*engine.Store, 0, len(s.registry))
-	for _, share := range s.registry {
-		if share.BlockStore != nil {
-			blockStores = append(blockStores, share.BlockStore)
-		}
-	}
-	s.mu.RUnlock()
-
-	var total int64
-	for _, bs := range blockStores {
-		completed, failed := bs.SyncCounts()
-		total += int64(completed) + int64(failed)
-	}
-	return total
-}
-
-// ShareBlockStoreStats holds block store statistics for a single share.
-type ShareBlockStoreStats struct {
-	ShareName string                 `json:"share_name"`
-	Stats     engine.BlockStoreStats `json:"stats"`
-}
-
-// BlockStoreStatsResponse holds aggregated and per-share block store statistics.
-type BlockStoreStatsResponse struct {
-	Totals   engine.BlockStoreStats `json:"totals"`
-	PerShare []ShareBlockStoreStats `json:"per_share,omitempty"`
-}
-
-// EvictOptions controls which block store tiers to evict.
-type EvictOptions struct {
-	ReadBufferOnly bool `json:"read_buffer_only"`
-	LocalOnly      bool `json:"local_only"`
-}
-
-// EvictResult holds the result of a block store eviction operation.
-type EvictResult struct {
-	ReadBufferEntriesCleared int   `json:"read_buffer_entries_cleared"`
-	LocalFilesEvicted        int   `json:"local_files_evicted"`
-	BytesFreed               int64 `json:"bytes_freed"`
-}
-
-// MetricsBlockStats returns per-share block-store stats for observability. It
-// uses the lite stats path (no per-file block-count DB walk) so a metrics
-// scrape never pays for the per-block-state counts, which are not exported.
-func (s *Service) MetricsBlockStats() []ShareBlockStoreStats {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	var out []ShareBlockStoreStats
-	for name, share := range s.registry {
-		if share.BlockStore == nil {
-			continue
-		}
-		out = append(out, ShareBlockStoreStats{ShareName: name, Stats: share.BlockStore.GetStatsLite()})
-	}
-	return out
-}
-
-// GetBlockStoreStats returns block store statistics, optionally filtered by share name.
-// If shareName is empty, returns aggregated stats across all shares with per-share breakdown.
-func (s *Service) GetBlockStoreStats(shareName string) (*BlockStoreStatsResponse, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	if shareName != "" {
-		share, exists := s.registry[shareName]
-		if !exists {
-			return nil, fmt.Errorf("share %q not found", shareName)
-		}
-		if share.BlockStore == nil {
-			return nil, fmt.Errorf("share %q has no block store configured", shareName)
-		}
-		stats := share.BlockStore.GetStats()
-		return &BlockStoreStatsResponse{
-			Totals:   stats,
-			PerShare: []ShareBlockStoreStats{{ShareName: shareName, Stats: stats}},
-		}, nil
-	}
-
-	var totals engine.BlockStoreStats
-	var perShare []ShareBlockStoreStats
-
-	for name, share := range s.registry {
-		if share.BlockStore == nil {
-			continue
-		}
-		stats := share.BlockStore.GetStats()
-		perShare = append(perShare, ShareBlockStoreStats{ShareName: name, Stats: stats})
-		addBlockStoreStats(&totals, stats)
-	}
-
-	return &BlockStoreStatsResponse{
-		Totals:   totals,
-		PerShare: perShare,
-	}, nil
-}
-
-// addBlockStoreStats accumulates src into dst (field-by-field summation).
-func addBlockStoreStats(dst *engine.BlockStoreStats, src engine.BlockStoreStats) {
-	dst.FileCount += src.FileCount
-	dst.BlocksDirty += src.BlocksDirty
-	dst.BlocksLocal += src.BlocksLocal
-	dst.BlocksRemote += src.BlocksRemote
-	dst.BlocksTotal += src.BlocksTotal
-	dst.BlocksCached += src.BlocksCached
-	dst.LocalDiskUsed += src.LocalDiskUsed
-	dst.LocalDiskMax += src.LocalDiskMax
-	dst.LocalMemUsed += src.LocalMemUsed
-	dst.LocalMemMax += src.LocalMemMax
-	dst.AppendLogLimitBytes += src.AppendLogLimitBytes
-	dst.ReadBufferEntries += src.ReadBufferEntries
-	dst.ReadBufferUsed += src.ReadBufferUsed
-	dst.ReadBufferMax += src.ReadBufferMax
-	dst.PendingSyncs += src.PendingSyncs
-	dst.PendingUploads += src.PendingUploads
-	dst.CompletedSyncs += src.CompletedSyncs
-	dst.FailedSyncs += src.FailedSyncs
-	dst.UnsyncedBytes += src.UnsyncedBytes
-	dst.OfflineReadsBlocked += src.OfflineReadsBlocked
-	if src.HasRemote {
-		dst.HasRemote = true
-	}
-}
-
-// EvictBlockStore evicts block store data for the given share (or all shares if shareName is empty).
-// Returns an error if trying to evict local blocks without a remote store (safety check).
-func (s *Service) EvictBlockStore(ctx context.Context, shareName string, opts EvictOptions) (*EvictResult, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	var targets []*Share
-	if shareName != "" {
-		share, exists := s.registry[shareName]
-		if !exists {
-			return nil, fmt.Errorf("share %q not found", shareName)
-		}
-		if share.BlockStore == nil {
-			return nil, fmt.Errorf("share %q has no block store configured", shareName)
-		}
-		targets = []*Share{share}
-	} else {
-		for _, share := range s.registry {
-			if share.BlockStore != nil {
-				targets = append(targets, share)
-			}
-		}
-	}
-
-	var result EvictResult
-
-	for _, share := range targets {
-		bs := share.BlockStore
-
-		if !opts.ReadBufferOnly && !bs.HasRemoteStore() {
-			return nil, fmt.Errorf("cannot evict local blocks for share %q: no remote store configured (data would be lost)", share.Name)
-		}
-
-		if !opts.LocalOnly {
-			result.ReadBufferEntriesCleared += bs.DestroyCache()
-		}
-
-		if !opts.ReadBufferOnly {
-			beforeDisk := bs.LocalStats().DiskUsed
-
-			files := bs.ListFiles()
-			for _, payloadID := range files {
-				_ = bs.EvictLocal(ctx, payloadID)
-				result.LocalFilesEvicted++
-			}
-
-			// EvictLocal only clears per-file append-log/memory state, and
-			// ListFiles goes empty after rollup — so post-rollup the resident
-			// bytes live in sealed log blobs it never touches. Drain them now
-			// (synced-only; safe because the no-remote refusal above guarantees
-			// a remote copy exists) so reads fall back to the remote.
-			drained, err := bs.DrainLocalSynced(ctx)
-			if err != nil {
-				return nil, fmt.Errorf("drain local synced blocks for share %q: %w", share.Name, err)
-			}
-
-			// Report the larger of the observed DiskUsed delta (captures the
-			// EvictLocal loop plus the drain) and the drain's own freed count.
-			// The raw delta can go negative if a concurrent write grows DiskUsed
-			// mid-eviction; max with the non-negative drained count keeps
-			// BytesFreed honest and never negative.
-			result.BytesFreed += max(beforeDisk-bs.LocalStats().DiskUsed, drained)
-		}
-	}
-
-	return &result, nil
-}
-
-// StartWarm starts (or returns the already-running) async warm job for
-// shareName. The job proactively materializes every remote block of the share
-// onto the local tier (engine.Store.WarmAll) on a DETACHED context so it
-// outlives the request that triggered it; RemoveShare cancels it. Returns an
-// error if the share is unknown or has no block store. The returned WarmJob is
-// a snapshot taken under the registry lock.
-func (s *Service) StartWarm(_ context.Context, shareName string) (*WarmJob, error) {
-	s.mu.RLock()
-	share, exists := s.registry[shareName]
-	if !exists {
-		s.mu.RUnlock()
-		return nil, fmt.Errorf("%w: %q", ErrShareNotFound, shareName)
-	}
-	if share.BlockStore == nil {
-		s.mu.RUnlock()
-		return nil, fmt.Errorf("share %q has no block store configured", shareName)
-	}
-	bs := share.BlockStore
-	s.mu.RUnlock()
-
-	// Capture the share's currently-stored byte count up front (O(1) lite
-	// stats, no per-file DB walk). The registry uses it to warn when a warm
-	// run enumerates zero blocks on a non-empty share (#1374).
-	usedBytes := bs.GetStatsLite().LocalDiskUsed
-
-	job := s.warmJobs.start(shareName, usedBytes, func(ctx context.Context, progress func(done, total int64)) (warmAllResult, error) {
-		res, err := bs.WarmAll(ctx, progress)
-		return warmAllResult{
-			BlocksFetched: res.BlocksFetched,
-			BytesFetched:  res.BytesFetched,
-		}, err
-	})
-	return job, nil
-}
-
-// GetWarm returns a snapshot of the warm job by ID, or (nil, false) if unknown.
-func (s *Service) GetWarm(jobID string) (*WarmJob, bool) {
-	return s.warmJobs.get(jobID)
-}
-
-// deriveGCStateRoot returns the per-share gc-state directory used by the GC
-// engine to persist its run state and last-run.json: the share's local store
-// directory plus a `gc-state` suffix. Returns "" whenever that directory is
-// unresolvable — engine.PersistLastRunSummary treats "" as "do not persist".
-func deriveGCStateRoot(localCfg interface {
-	GetConfig() (map[string]any, error)
-}, shareName string) string {
-	dir := deriveLocalStoreDir(localCfg, shareName)
-	if dir == "" {
-		return ""
-	}
-	return filepath.Join(dir, "gc-state")
-}
-
-// deriveLocalStoreDir returns the per-share on-disk data directory the
-// migration tool uses to host `.migration-state.jsonl` and the rolling
-// snapshot. Returns "" for in-memory or unresolvable configs (the REST
-// status handler treats "" as "no journal available", not an error).
-//
-// Path layout: `<basePath>/shares/<sanitized>/`. Note the absence of a
-// "blocks" or "gc-state" suffix — the migration journal lives at the
-// share root next to the blocks directory, not inside it. This matches
-// the offline migration tool's --state-dir contract: the operator
-// passes the same path the daemon would compute here.
-func deriveLocalStoreDir(localCfg interface {
-	GetConfig() (map[string]any, error)
-}, shareName string) string {
-	if localCfg == nil {
-		return ""
-	}
-	cfg, err := localCfg.GetConfig()
-	if err != nil {
-		return ""
-	}
-	basePath, ok := cfg["path"].(string)
-	if !ok || basePath == "" {
-		return ""
-	}
-	expanded, err := pathutil.ExpandPath(basePath)
-	if err != nil {
-		return ""
-	}
-	if !filepath.IsAbs(expanded) {
-		return ""
-	}
-	return filepath.Join(expanded, "shares", sanitizeShareName(shareName))
-}
-
-const (
-	// minDirtyExpire is the shortest dirty-age commit interval a share may
-	// configure. Anything below it is a misconfiguration rather than a tuning
-	// choice: the loop would issue barriers faster than a disk retires them.
-	minDirtyExpire = time.Second
-	// maxDirtyExpireSeconds is where a seconds value stops fitting a
-	// time.Duration (~292 years).
-	maxDirtyExpireSeconds = float64(math.MaxInt64 / int64(time.Second))
-)
-
-// dirtyExpiryFromConfig reads the dirty_expire_seconds key, which caps how long
-// an acknowledged write may stay in the page cache before the journal fsyncs it
-// on its own. Zero (absent, or an unusable value) leaves the journal default in
-// place; a negative value disables the loop, leaving the client's own fsync and
-// segment rotation as the only durability points.
-func dirtyExpiryFromConfig(config map[string]any) time.Duration {
-	v, ok := config["dirty_expire_seconds"]
-	if !ok {
-		return 0
-	}
-	n, isNum := v.(float64)
-	if !isNum || math.IsNaN(n) || math.Abs(n) > maxDirtyExpireSeconds {
-		logger.Warn("block store config has dirty_expire_seconds but it is invalid; ignoring", "value", v)
-		return 0
-	}
-	d := time.Duration(n * float64(time.Second))
-	// A sub-second interval would put a disk barrier on the store far more often
-	// than it can retire one; a typo must not do that.
-	if n > 0 && d < minDirtyExpire {
-		logger.Warn("block store config dirty_expire_seconds is below the floor; clamping",
-			"value", n, "floor", minDirtyExpire)
-		return minDirtyExpire
-	}
-	return d
-}
-
-// CreateLocalStoreFromConfig creates a local store instance from a block store config.
-func CreateLocalStoreFromConfig(
-	ctx context.Context,
-	storeType string,
-	cfg interface {
-		GetConfig() (map[string]any, error)
-	},
-	shareName string,
-	defaults *LocalStoreDefaults,
-	fileChunkStore block.EngineFileChunkStore,
-	migrateLegacy bool,
-) (local.LocalStore, error) {
-	config, err := cfg.GetConfig()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get config: %w", err)
-	}
-
-	var maxDisk int64
-	if defaults != nil {
-		maxDisk = int64(defaults.MaxSize)
-	}
-
-	// Remote-cache backpressure window (how long a write stalls for the
-	// syncer to drain before ErrDiskFull). Threaded into FSStoreOptions
-	// below; zero defers to the FSStore default.
-	var backpressureMaxWait time.Duration
-	if defaults != nil {
-		backpressureMaxWait = defaults.BackpressureMaxWait
-	}
-
-	// Per-store max_size from config JSON takes precedence over defaults
-	if v, ok := config["max_size"]; ok {
-		if n, ok := v.(float64); ok && n > 0 {
-			maxDisk = int64(n)
-		} else {
-			logger.Warn("block store config has max_size but it is invalid or non-positive; ignoring", "value", v)
-		}
-	}
-
-	// Append is mandatory on the local tier — the use_append_log opt-out
-	// flag was deleted with the legacy path-keyed writer. Budgets still
-	// surface through FSStoreOptions to fs.NewWithOptions; invalid values
-	// are warned and ignored.
-	var fsOpts fs.FSStoreOptions
-	fsOpts.BackpressureMaxWait = backpressureMaxWait
-	// A remote-backed share may carry a pre-journal blobs/+logs/ layout from an
-	// upgrade; archive it aside so the journal opens clean (the caller then cold-
-	// seeds from the surviving manifest). A local-only share has no remote to
-	// re-fetch from, so it takes the async log-only migration path instead: the
-	// bytes are re-ingested from the surviving append logs when they are complete
-	// (no compacted log), and the guardrail stays fatal otherwise.
-	fsOpts.MigrateLegacyLayout = migrateLegacy
-	fsOpts.MigrateLegacyLocalOnly = !migrateLegacy
-	// Local-cache size-hint default. Precedence (lowest first):
-	// FSStore internal default < global/deduced default (plumbed via
-	// LocalStoreDefaults.MaxLogBytes) < per-store config["max_log_bytes"].
-	// max_log_bytes no longer gates writes; it only feeds the Stats size hint.
-	// Seed fsOpts.MaxLogBytes from the global/deduced default here; the
-	// per-store config branch below overrides it when present.
-	if defaults != nil && defaults.MaxLogBytes > 0 {
-		fsOpts.MaxLogBytes = defaults.MaxLogBytes
-	}
-	if v, ok := config["max_log_bytes"]; ok {
-		if n, ok := v.(float64); ok && n > 0 {
-			// FIX-15: JSON-decoded numbers land here as float64. Values above
-			// 2^53 (~9 PiB) lose integer precision, and non-integer values
-			// silently truncate. Warn so a misconfigured budget surfaces in
-			// logs instead of producing a budget that is off by hundreds of
-			// kilobytes from what the operator typed.
-			// Reject out-of-range and non-integer values rather than perform
-			// an implementation-defined float64->int64 cast (which on out-of-range
-			// inputs can produce a negative or garbage budget).
-			if n > float64(math.MaxInt64) || n != math.Trunc(n) {
-				logger.Warn("config: max_log_bytes is out of range or non-integer; keeping default", "value", n)
-			} else {
-				fsOpts.MaxLogBytes = int64(n)
-			}
-		} else {
-			logger.Warn("block store config has max_log_bytes but it is invalid or non-positive; ignoring", "value", v)
-		}
-	}
-	fsOpts.DirtyExpiry = dirtyExpiryFromConfig(config)
-	// chunk_size sets the FastCDC Min for this share's carve chunker (#1569) —
-	// the dominant knob for effective chunk size and thus random-read
-	// amplification. Avg/Max are derived (4x/8x Min) unless chunk_max overrides
-	// the ceiling. Absent => the FSStore default (1M/4M/16M, byte-identical to
-	// pre-#1569). Lower it (e.g. 131072 = 128 KiB) on random-access shares
-	// (VM images / databases): trades weaker dedup + more FileChunk manifest
-	// rows for far less read amplification. Reads never re-chunk, so changing
-	// this only affects newly written data.
-	if v, ok := config["chunk_size"]; ok {
-		if n, ok := v.(float64); ok && n > 0 && n == math.Trunc(n) && n <= float64(math.MaxInt32) {
-			cp := chunker.Params{Min: int(n), Avg: int(n) * 4, Max: int(n) * 8}
-			if mv, ok := config["chunk_max"]; ok {
-				if m, ok := mv.(float64); ok && m > 0 && m == math.Trunc(m) && m <= float64(math.MaxInt32) {
-					cp.Max = int(m)
-					if cp.Avg > cp.Max {
-						cp.Avg = cp.Max
-					}
-				} else {
-					logger.Warn("block store config has chunk_max but it is invalid; ignoring", "value", mv)
-				}
-			}
-			if err := cp.Validate(); err != nil {
-				logger.Warn("block store config chunk_size produced invalid chunker params; keeping default", "error", err)
-			} else {
-				fsOpts.ChunkParams = cp
-			}
-		} else {
-			logger.Warn("block store config has chunk_size but it is invalid or non-positive; ignoring", "value", v)
-		}
-	}
-	switch storeType {
-	case "fs":
-		basePath, ok := config["path"].(string)
-		if !ok || basePath == "" {
-			return nil, errors.New("fs local store requires path in config")
-		}
-		expanded, err := pathutil.ExpandPath(basePath)
-		if err != nil {
-			return nil, fmt.Errorf("failed to expand path %q: %w", basePath, err)
-		}
-		// Defense-in-depth: ValidateBlockStoreConfig rejects relative paths at
-		// create/update time, but pre-existing or out-of-band configs could
-		// still carry them. Guard here so filepath.Join doesn't resolve
-		// against the server's CWD.
-		if !filepath.IsAbs(expanded) {
-			return nil, fmt.Errorf("fs local store path must be absolute, got %q", basePath)
-		}
-		sanitized := sanitizeShareName(shareName)
-		// The FSStore creates `blocks/` (CAS) and `logs/` (append log) as
-		// siblings under its baseDir. A previous layout produced a doubled
-		// `shares/{name}/blocks/blocks/...` path. Existing pre-v0.16 installs
-		// migrate via `dfs migrate-to-cas` (which uses share-root as its
-		// state-dir, already aligned with deriveLocalStoreDir).
-		shareDir := filepath.Join(expanded, "shares", sanitized)
-		if err := os.MkdirAll(shareDir, 0755); err != nil {
-			return nil, fmt.Errorf("failed to create share directory: %w", err)
-		}
-
-		// The SyncedHashStore and LocalChunkIndex are derived from this same
-		// fileChunkStore backend inside NewWithOptions. There is no longer a
-		// rollup worker pool (the journal carves dirty ranges directly), so the
-		// former RollupStore guard and StartRollup call are gone.
-		store, err := fs.NewWithOptions(shareDir, maxDisk, fileChunkStore, fsOpts)
-		if err != nil {
-			return nil, err
-		}
-		applyDurableOverride(store, config, "local "+storeType, shareName)
-		return store, nil
-
-	case "memory":
-		store := localmemory.New()
-		applyDurableOverride(store, config, "local "+storeType, shareName)
-		return store, nil
-
-	default:
-		return nil, fmt.Errorf("unsupported local store type: %s", storeType)
-	}
-}
-
-// durableOverrideSetter is implemented by block stores (local and remote) that
-// expose a per-store durability override (block stores embed a
-// block.DurabilityReporter type-default; this lets an operator flip it).
-type durableOverrideSetter interface {
-	SetDurable(bool)
-}
-
-// applyDurableOverride reads an optional "durable" bool from the per-store
-// config and applies it to store when present, overriding the type-default
-// durability (#1274). A non-bool "durable" value is warned and ignored so the
-// type-default stands. label/shareName feed the diagnostic log line.
-func applyDurableOverride(store any, config map[string]any, label, shareName string) {
-	v, ok := config["durable"]
-	if !ok {
-		return
-	}
-	b, ok := v.(bool)
-	if !ok {
-		logger.Warn("block store config has durable but it is not a bool; ignoring",
-			"store", label, "share", shareName, "value", v)
-		return
-	}
-	setter, ok := store.(durableOverrideSetter)
-	if !ok {
-		// Every shipped store implements SetDurable; a store that does not is a
-		// programmer error, surface it instead of silently dropping the override.
-		logger.Warn("block store does not support a durable override; ignoring config[\"durable\"]",
-			"store", label, "share", shareName)
-		return
-	}
-	setter.SetDurable(b)
-	logger.Info("block store durability overridden by config", "store", label, "share", shareName, "durable", b)
-}
-
-// SeedColdFromManifest seeds a cold journal interval for every FileChunk in the
-// share's metadata manifest so a subsequent read faults the bytes in from the
-// remote store instead of zero-filling. It is used when the journal has no local
-// copy of the data — after a snapshot restore wiped the local tier, or after a
-// pre-journal upgrade archived the legacy local layout aside. Remote-backed
-// shares only (the caller gates on that); the cold fetch it arms is
-// BLAKE3-verified. One ListFileChunks and one SeedCold per payload — O(chunks),
-// acceptable for a rare control-plane / startup path, and seeding a whole file at
-// once keeps the local tier's durable write to one per file.
-//
-// The report it returns is what the seed observed on the way through: how much
-// it covered, how much of it the manifest does not yet call remote, and a few
-// extents to read back. Seeding alone proves nothing about content, so a caller
-// that archived the only local copy aside is expected to verify before it
-// reports success.
-func SeedColdFromManifest(ctx context.Context, bs *engine.Store, metaStore metadata.Store) (coldSeedReport, error) {
-	var report coldSeedReport
-	// EnumeratePayloads is a callback iteration with no cheap denominator, so
-	// the heartbeat reports a running count rather than a fraction.
-	started := time.Now()
-	lastLog := started
-	logger.Info("seeding cold intervals from the metadata manifest")
-	err := metaStore.EnumeratePayloads(ctx, func(payloadID string) error {
-		rows, err := metaStore.ListFileChunks(ctx, payloadID)
-		if err != nil {
-			return fmt.Errorf("list manifest for %s: %w", payloadID, err)
-		}
-		extents := make([][2]int64, 0, len(rows))
-		for _, row := range rows {
-			if row == nil {
-				continue
-			}
-			off, ok := block.ParseChunkOffset(row.ID)
-			if !ok {
-				// Which range this row describes cannot be recovered — that is what
-				// makes it unplaceable — so it is logged verbatim and seeding
-				// continues: one such row is not worth stranding every other share.
-				// Leaving the range unseeded is safe because the engine reconciles an
-				// uncovered read against the manifest whether or not the range is
-				// cold, so it refuses with ErrManifestInconsistent rather than serving
-				// zeros.
-				logger.Error("cold seed: unplaceable manifest row, range will not be seeded",
-					"payload", payloadID, "row", row.ID, "size", row.DataSize)
-				report.unplaceable++
-				continue
-			}
-			extents = append(extents, [2]int64{int64(off), int64(row.DataSize)})
-			report.chunks++
-			// Whether a chunk reached the remote is answered by the synced-hash
-			// store, not by FileChunk.State: the carve path records synced markers
-			// and leaves the row state at Pending for the life of the payload, so
-			// reading State here would call every chunk unsynced. A lookup failure
-			// counts as unsynced — the archive stays when we cannot tell.
-			synced, serr := metaStore.IsSynced(ctx, row.Hash)
-			if serr != nil || !synced {
-				report.unsynced++
-			}
-			// Sample the first hashed extent of the first few payloads. A
-			// zero-length chunk or one with no hash yet cannot be checked
-			// against anything, so it is not worth a sample slot.
-			if len(report.samples) < coldVerifySamples && row.DataSize > 0 &&
-				row.Hash != (block.ContentHash{}) && report.sampledPayload(payloadID) {
-				report.samples = append(report.samples, coldSample{
-					payloadID: payloadID,
-					offset:    int64(off),
-					length:    int64(row.DataSize),
-					hash:      row.Hash,
-				})
-			}
-		}
-		if err := bs.SeedCold(ctx, payloadID, extents); err != nil {
-			return fmt.Errorf("seed cold %s: %w", payloadID, err)
-		}
-		report.payloads++
-		if time.Since(lastLog) >= migrationProgressInterval {
-			lastLog = time.Now()
-			logger.Info("seeding cold intervals from the metadata manifest",
-				"payloads", report.payloads, "chunks", report.chunks,
-				"elapsed", time.Since(started).Round(time.Second))
-		}
-		return nil
-	})
-	return report, err
-}
-
-// parseRequireDurableCommit reads the optional per-share "require_durable_commit"
-// bool from the local store config (#1274). Read the same conservative way as
-// config["durable"]: absent or non-bool → false (default), so the commit seam
-// acks once Flush succeeds and the remote mirror stays async — ordinary
-// NFS/POSIX writes never EIO. When true, CLOSE/COMMIT only succeed once the data
-// is on a durable store.
-func parseRequireDurableCommit(config map[string]any, shareName string) bool {
-	v, ok := config["require_durable_commit"]
-	if !ok {
-		return false
-	}
-	b, ok := v.(bool)
-	if !ok {
-		logger.Warn("block store config has require_durable_commit but it is not a bool; ignoring",
-			"share", shareName, "value", v)
-		return false
-	}
-	return b
-}
-
-// resolveDurabilityTier composes the per-share durability knobs into the two
-// underlying behaviors (#1758). The optional "durability" enum in the local
-// store config selects a named tier; when absent, the older raw bools
-// ("writeback", "require_durable_commit") are honored unchanged for backward
-// compatibility. When "durability" is present it is authoritative — the raw
-// bools are ignored.
-//
-//	local     (default) — journal + metadata fsync, async S3 (node-crash-safe)
-//	writeback           — per-op FILE_SYNC metadata flush relaxed (deferred to
-//	                      the ticker); data still journal-fsync durable. The full
-//	                      data-writeback tier additionally needs the journal
-//	                      async-commit half, tracked separately.
-//	remote              — CLOSE/COMMIT block until data is durable in the remote
-//	                      (S3) store (require_durable_commit); survives node loss.
-func resolveDurabilityTier(config map[string]any, shareName string) (writeback, requireDurableCommit bool) {
-	v, ok := config["durability"]
-	if !ok {
-		return parseWritebackConfig(config, shareName), parseRequireDurableCommit(config, shareName)
-	}
-	tier, ok := v.(string)
-	if !ok {
-		logger.Warn("block store config has durability but it is not a string; defaulting to local",
-			"share", shareName, "value", v)
-		return false, false
-	}
-	switch strings.ToLower(strings.TrimSpace(tier)) {
-	case "", "local":
-		return false, false
-	case "writeback":
-		logger.Info("durability tier: writeback (metadata flush relaxed)", "share", shareName)
-		return true, false
-	case "remote":
-		logger.Info("durability tier: remote (ack-on-S3, strict CLOSE/COMMIT)", "share", shareName)
-		return false, true
-	default:
-		logger.Warn("unknown durability tier; defaulting to local",
-			"share", shareName, "durability", tier)
-		return false, false
-	}
-}
-
-// parseWritebackConfig reads the optional per-share "writeback" bool from the
-// local store config (#1757). Read the same conservative way as
-// config["require_durable_commit"]: absent or non-bool → false (default,
-// durable). When true, the share's per-op FILE_SYNC metadata flush takes the
-// relaxed deferred-fsync path (see metadata.Service.SetShareWriteback).
-func parseWritebackConfig(config map[string]any, shareName string) bool {
-	v, ok := config["writeback"]
-	if !ok {
-		return false
-	}
-	b, ok := v.(bool)
-	if !ok {
-		logger.Warn("block store config has writeback but it is not a bool; ignoring",
-			"share", shareName, "value", v)
-		return false
-	}
-	if b {
-		logger.Info("metadata writeback tier enabled by config", "share", shareName)
-	}
-	return b
-}
-
-// CreateRemoteStoreFromConfig creates a remote store from type and dynamic config.
-func CreateRemoteStoreFromConfig(ctx context.Context, storeType string, cfg interface {
-	GetConfig() (map[string]any, error)
-}) (remote.RemoteStore, error) {
-	config, err := cfg.GetConfig()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get config: %w", err)
-	}
-
-	switch storeType {
-	case "memory":
-		store := remotememory.New()
-		applyDurableOverride(store, config, "remote "+storeType, "")
-		return store, nil
-
-	case "filesystem":
-		return nil, errors.New("remote store type 'filesystem' removed in v4.0 -- use 'memory' or 's3'")
-
-	case "s3":
-		bucket, ok := config["bucket"].(string)
-		if !ok || bucket == "" {
-			return nil, errors.New("s3 remote store requires bucket")
-		}
-
-		region := "us-east-1"
-		if r, ok := config["region"].(string); ok && r != "" {
-			region = r
-		}
-
-		endpoint, _ := config["endpoint"].(string)
-		prefix, _ := config["prefix"].(string)
-		accessKey, _ := config["access_key_id"].(string)
-		secretKey, _ := config["secret_access_key"].(string)
-		if accessKey == "" || secretKey == "" {
-			return nil, errors.New("s3 remote store requires access_key_id and secret_access_key")
-		}
-		// When a custom endpoint is set (MinIO, Synology, etc.), default to
-		// path-style addressing — virtual-hosted style rarely works on
-		// non-AWS S3-compatible services. This matches v0.8.x behavior.
-		// Only override when the key is absent; honor explicit false.
-		forcePathStyle, hasPathStyle := config["force_path_style"].(bool)
-		if endpoint != "" && !hasPathStyle {
-			forcePathStyle = true
-		}
-		allowPrivate, _ := config["allow_private_endpoint"].(bool)
-
-		store, err := remotes3.NewFromConfig(ctx, remotes3.Config{
-			Bucket:         bucket,
-			Region:         region,
-			Endpoint:       endpoint,
-			AccessKey:      accessKey,
-			SecretKey:      secretKey,
-			KeyPrefix:      prefix,
-			ForcePathStyle: forcePathStyle,
-			AllowPrivate:   allowPrivate,
-		})
-		if err != nil {
-			return nil, err
-		}
-		applyDurableOverride(store, config, "remote "+storeType, "")
-		return store, nil
-
-	default:
-		return nil, fmt.Errorf("unsupported remote store type: %s", storeType)
 	}
 }
